refactor: extract timeline rendering
