+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -1897,7 +1897,7 @@ private fun StartupDataOverlay() {
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-private fun KgsModalBottomSheet(
+internal fun KgsModalBottomSheet(
     onDismissRequest: () -> Unit,
     modifier: Modifier = Modifier,
     containerColor: Color = MaterialTheme.colorScheme.surface,
@@ -4057,7 +4057,7 @@ private fun CalendarShell(
           val morphScope = this
           CompositionLocalProvider(LocalMorphAnimatedVisibilityScope provides morphScope) {
             when (group) {
-                CalendarViewGroup.Timeline -> ThreeDayView(
+                CalendarViewGroup.Timeline -> TimelineView(
                     state = state,
                     selectedView = state.selectedView,
                     onDateSelected = onDateSelected,
@@ -5434,833 +5434,6 @@ private fun YearStrip(
     }
 }
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun ThreeDayView(
-    state: CalendarUiState,
-    // The active timeline view — Day (1 column) or ThreeDay (3 columns). The same composable
-    // serves both; switching between them animates the column width rather than swapping.
-    selectedView: CalendarViewMode,
-    onDateSelected: (LocalDate) -> Unit,
-    onViewSelected: (CalendarViewMode) -> Unit,
-    onMultiDayCountChanged: (Int) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEventMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onTaskMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onEventMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onTaskMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onSlotSelected: (LocalDate, LocalTime) -> Unit,
-    onAllDaySlotSelected: (LocalDate) -> Unit,
-    draftEvent: DraftEventSelection?,
-    onDraftEventChanged: (DraftEventSelection) -> Unit,
-    onDraftInteraction: () -> Unit,
-    onDraftTap: () -> Unit,
-    timelineBottomInset: Dp,
-    onDetail: (DetailSheet) -> Unit,
-    // Hoisted so the vertical zoom (pinch) and scroll position are shared between the
-    // 3-day and 1-day instances and survive the morph between them.
-    timeScroll: androidx.compose.foundation.ScrollState,
-    hourHeightDp: Float,
-    onHourHeightChange: (Float) -> Unit,
-    initialTimeScrollApplied: Boolean,
-    onInitialTimeScrollApplied: () -> Unit,
-    monthMorphDay: LocalDate,
-) {
-    val scope = rememberCoroutineScope()
-    val calendarTime = LocalCalendarTimeSnapshot.current
-    val now = calendarTime.nowMinute
-    val density = LocalDensity.current
-    var dayViewportWidthPx by remember { mutableStateOf(0) }
-    var gridViewportHeightPx by remember { mutableStateOf(0) }
-    val bodyScrollableState = rememberScrollableState { delta ->
-        val consumed = timeScroll.dispatchRawDelta(-delta)
-        -consumed
-    }
-    val daySpacingPx = with(density) { DayColumnSpacing.toPx() }
-    // The tapped day that is currently expanding into / collapsing out of the 1-day view, plus
-    // the 3-day window it morphs within. Non-null only while a header-tap morph is playing; it
-    // drives the in-place "zoom around the tapped column" overlay below. Cleared when the
-    // column-count animation settles.
-    var morphContext by remember { mutableStateOf<DayMorphContext?>(null) }
-    // The timeline's OWN mode (Day vs ThreeDay), held independently of selectedView. When the app
-    // switches away to month/agenda/tasks, the timeline keeps its last mode while it exits — so a
-    // 1-day view collapses back into a month cell (and a 3-day view does NOT pretend to be a day).
-    // selectedView is momentarily "Month" for the outgoing timeline, which would otherwise flip
-    // this to 3-day and break the reverse morph.
-    var timelineMode by remember {
-        mutableStateOf(if (selectedView == CalendarViewMode.Day) CalendarViewMode.Day else CalendarViewMode.ThreeDay)
-    }
-    LaunchedEffect(selectedView) {
-        if (selectedView == CalendarViewMode.Day) timelineMode = CalendarViewMode.Day
-        else if (selectedView == CalendarViewMode.ThreeDay) timelineMode = CalendarViewMode.ThreeDay
-    }
-    val isDayMode = timelineMode == CalendarViewMode.Day
-    val multiDayCount = state.multiDayCount.coerceMultiDayCount()
-    val targetDayCount = if (isDayMode) 1 else multiDayCount
-    val dayEndGutterPx = daySpacingPx
-    val dayGeometryViewportWidthPx = (dayViewportWidthPx - dayEndGutterPx).coerceAtLeast(0f)
-    // Animate the column COUNT as a float between 1 and the configured multi-day count. The pager's fixed page size is
-    // derived from it for the steady states and the back-button morph; the header-tap morph
-    // instead uses [morphContext] + the overlay so the tapped day can expand from ANY column
-    // (the pager can only ever anchor the leftmost page).
-    val animatedDayCount by animateFloatAsState(
-        targetValue = targetDayCount.toFloat(),
-        animationSpec = tween(MorphDurationMs, easing = MorphEasing),
-        label = "dayColumnCount",
-    )
-    // 0 = full multi-day, 1 = full 1-day. Derived from the animating count so the overlay and the
-    // underlying pager are always perfectly in step.
-    val morphSpan = (multiDayCount - 1).coerceAtLeast(1).toFloat()
-    val morphProgress = ((multiDayCount - animatedDayCount) / morphSpan).coerceIn(0f, 1f)
-    // Day indexing (which dates are visible, all-day stacking) uses the integer target so the
-    // content is correct for where we're heading; only the WIDTH interpolates.
-    // Track the in-between count as the width animates (round up) so the visible-day window and
-    // the all-day band reflect the slide — e.g. a neighbour day's all-day chips stay rendered
-    // while it slides out, instead of snapping to the destination count the instant the morph
-    // starts.
-    val clampedDayCount = ceil(animatedDayCount).roundToInt().coerceIn(1, MAX_MULTI_DAY_COUNT)
-    val dayWidthPx = ((dayGeometryViewportWidthPx - ((animatedDayCount - 1f) * daySpacingPx)) / animatedDayCount)
-        .coerceAtLeast(with(density) { 32.dp.toPx() })
-    val dayWidthDp = with(density) { dayWidthPx.toDp() }
-    // Anchor the pager on selectedDate itself, NOT on visibleRange.startDate. The loaded range
-    // is intentionally wider than the shown day (for the morphs), so its start is several days
-    // earlier; using it here would put the pager — and the settle feedback below — on the wrong
-    // day and send it scrolling.
-    val selectedPage = state.selectedDate.toDayPage()
-    val pagerState = rememberPagerState(initialPage = selectedPage, pageCount = { DayPagerPageCount })
-    var programmaticTargetPage by remember { mutableStateOf<Int?>(null) }
-    var handledWidgetDateNavigationSerial by remember { mutableStateOf(0) }
-    val latestMorphContext by rememberUpdatedState(morphContext)
-    val latestSelectedDate by rememberUpdatedState(state.selectedDate)
-    val actualVisiblePages = remember(
-        pagerState.layoutInfo.visiblePagesInfo,
-        dayViewportWidthPx,
-        dayGeometryViewportWidthPx,
-        dayWidthPx,
-        pagerState.currentPage,
-        clampedDayCount,
-    ) {
-        val viewportWidth = dayViewportWidthPx
-        val geometryViewportWidth = dayGeometryViewportWidthPx
-        val pageExtent = dayWidthPx + daySpacingPx
-        pagerState.layoutInfo.visiblePagesInfo
-            .filter { page ->
-                viewportWidth <= 0 || (page.offset + pageExtent > 0 && page.offset < geometryViewportWidth + daySpacingPx)
-            }
-            .map { it.index.coerceIn(0, DayPagerPageCount - 1) }
-            .distinct()
-            .sorted()
-            .ifEmpty {
-                val start = pagerState.currentPage.coerceIn(0, DayPagerPageCount - 1)
-                (0 until clampedDayCount).map { (start + it).coerceIn(0, DayPagerPageCount - 1) }
-            }
-    }
-    val visibleStartDate = actualVisiblePages.minOrNull()?.toDayDate() ?: state.selectedDate
-    val visibleEndDate = actualVisiblePages.maxOrNull()?.toDayDate() ?: visibleStartDate.plusDays((clampedDayCount - 1).toLong())
-    val pagerVisibleDays = actualVisiblePages.map { it.toDayDate() }
-    val eventsByDay = remember(state.events) { state.events.indexEventsByDay() }
-    val calendarTasks = remember(state.datedTasks, state.showCompletedTasksInCalendar) {
-        if (state.showCompletedTasksInCalendar) state.datedTasks else state.datedTasks.filterNot { it.isCompleted }
-    }
-    val tasksByDay = remember(calendarTasks) { calendarTasks.indexTasksByDay() }
-    var allDayExpanded by remember { mutableStateOf(false) }
-    val allDayHeight = remember(pagerVisibleDays, state.events, calendarTasks, state.maxVisibleAllDayItems, allDayExpanded, draftEvent) {
-        pagerVisibleDays.allDayAreaHeight(
-            events = state.events,
-            tasks = calendarTasks,
-            maxVisibleItems = state.maxVisibleAllDayItems,
-            expanded = allDayExpanded,
-            draftDate = draftEvent?.takeIf { it.allDay }?.date,
-        )
-    }
-    val allDayHasOverflow = remember(pagerVisibleDays, state.events, calendarTasks, state.maxVisibleAllDayItems) {
-        pagerVisibleDays.hasAllDayOverflow(state.events, calendarTasks, state.maxVisibleAllDayItems)
-    }
-    val allDayArrowRotation by animateFloatAsState(
-        targetValue = if (allDayExpanded) 180f else 0f,
-        animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-        label = "allDayArrowRotation",
-    )
-    // Use the morph duration so the all-day band grows/shrinks in step with the column-width
-    // slide when toggling 1-day <-> 3-day.
-    val animatedAllDayHeight by animateDpAsState(
-        targetValue = allDayHeight,
-        animationSpec = tween(MorphDurationMs, easing = MorphEasing),
-        label = "allDayHeight",
-    )
-    val minHourHeightDp = if (gridViewportHeightPx > 0) {
-        max(AbsoluteMinHourRowHeightDp, with(density) { gridViewportHeightPx.toDp().value } / (DayEndHour - DayStartHour + 1))
-    } else {
-        AbsoluteMinHourRowHeightDp
-    }
-
-    LaunchedEffect(minHourHeightDp) {
-        if (hourHeightDp < minHourHeightDp) onHourHeightChange(minHourHeightDp)
-    }
-
-    LaunchedEffect(selectedPage, morphContext, state.dateNavigationSerial) {
-        // A header-tap morph owns the visible timeline until its final handoff. Moving the pager
-        // while that animation is starting makes a middle/right tapped column jump left before it
-        // expands. The handoff effect below settles the still-hidden pager immediately before the
-        // overlay is removed.
-        if (morphContext != null) return@LaunchedEffect
-        if (state.dateNavigationSerial > handledWidgetDateNavigationSerial) {
-            handledWidgetDateNavigationSerial = state.dateNavigationSerial
-            programmaticTargetPage = selectedPage
-            pagerState.scrollToPage(selectedPage)
-            programmaticTargetPage = null
-            return@LaunchedEffect
-        }
-        if (selectedView == CalendarViewMode.Day &&
-            pagerState.currentPage == selectedPage &&
-            abs(pagerState.currentPageOffsetFraction) > 0.001f
-        ) {
-            programmaticTargetPage = selectedPage
-            pagerState.scrollToPage(selectedPage)
-            programmaticTargetPage = null
-            return@LaunchedEffect
-        }
-        if (pagerState.currentPage != selectedPage && pagerState.settledPage != selectedPage) {
-            programmaticTargetPage = selectedPage
-            val distance = selectedPage - pagerState.currentPage
-            if (abs(distance) > 8) {
-                pagerState.scrollToPage(selectedPage - distance.coerceIn(-3, 3))
-            }
-            pagerState.animateScrollToPage(
-                page = selectedPage,
-                animationSpec = tween(durationMillis = 320, easing = MotionEmphasized),
-            )
-            programmaticTargetPage = null
-        }
-    }
-
-    LaunchedEffect(morphContext, animatedDayCount, targetDayCount, selectedPage, state.selectedDate) {
-        val ctx = morphContext
-        if (ctx != null && timelineMode == ctx.targetMode && abs(animatedDayCount - targetDayCount) < 0.001f) {
-            // Keep the overlay visible through one settled pager frame. Clearing it immediately
-            // from animateFloatAsState's completion callback reveals stale pager geometry for a
-            // single frame when the tapped day was not the leftmost 3-day column.
-            val focusedDate = ctx.days[ctx.expandSlot]
-            val focusedPage = focusedDate.toDayPage()
-            programmaticTargetPage = focusedPage
-            pagerState.scrollToPage(focusedPage)
-            if (state.selectedDate != focusedDate) {
-                // Commit the selected date only after the visual morph has reached its target.
-                // Updating it at tap-time makes middle/right columns relocate to the pager's
-                // left edge before their expansion animation begins.
-                onDateSelected(focusedDate)
-                return@LaunchedEffect
-            }
-            withFrameNanos { }
-            morphContext = null
-            programmaticTargetPage = null
-        }
-    }
-
-    LaunchedEffect(gridViewportHeightPx, hourHeightDp, timeScroll.maxValue) {
-        if (!initialTimeScrollApplied && gridViewportHeightPx > 0 && timeScroll.maxValue > 0) {
-            val target = with(density) { ((9 - DayStartHour) * hourHeightDp).dp.roundToPx() }
-            timeScroll.scrollTo(target.coerceIn(0, timeScroll.maxValue))
-            onInitialTimeScrollApplied()
-        }
-    }
-    val applyTimelineZoom: (Float, Float) -> Unit = { zoom, centroidY ->
-        val oldHourHeightDp = hourHeightDp
-        val newHourHeightDp = (oldHourHeightDp * zoom).coerceIn(minHourHeightDp, MaxHourRowHeightDp)
-        if (newHourHeightDp != oldHourHeightDp) {
-            val oldRowPx = with(density) { oldHourHeightDp.dp.toPx() }.coerceAtLeast(1f)
-            val newRowPx = with(density) { newHourHeightDp.dp.toPx() }.coerceAtLeast(1f)
-            val contentYBefore = timeScroll.value + centroidY
-            val anchoredContentRatio = contentYBefore / oldRowPx
-            val maxScrollPx = max(0f, ((DayEndHour - DayStartHour + 1) * newRowPx) - gridViewportHeightPx)
-            val nextScroll = (anchoredContentRatio * newRowPx - centroidY).coerceIn(0f, maxScrollPx)
-            onHourHeightChange(newHourHeightDp)
-            scope.launch { timeScroll.scrollTo(nextScroll.roundToInt()) }
-        }
-    }
-    val currentApplyTimelineZoom = rememberUpdatedState(applyTimelineZoom)
-    val timelinePinchZoom = remember {
-        { zoom: Float, centroidY: Float -> currentApplyTimelineZoom.value(zoom, centroidY) }
-    }
-    var previousTimelineInsetPx by remember { mutableStateOf(0) }
-    LaunchedEffect(timelineBottomInset, gridViewportHeightPx) {
-        val insetPx = with(density) { timelineBottomInset.roundToPx() }
-        val delta = insetPx - previousTimelineInsetPx
-        if (delta != 0 && gridViewportHeightPx > 0 && timeScroll.maxValue > 0) {
-            withFrameNanos { }
-            timeScroll.animateScrollTo((timeScroll.value + delta).coerceIn(0, timeScroll.maxValue))
-        }
-        previousTimelineInsetPx = insetPx
-    }
-
-    LaunchedEffect(pagerState) {
-        snapshotFlow { pagerState.settledPage }
-            .distinctUntilChanged()
-            .collect { page ->
-                // Header morphs deliberately resize the pager columns. That internal relayout can
-                // change settledPage even though the user did not horizontally scroll. Treating it
-                // like a swipe is the legacy focus behavior that moves a tapped middle/right date
-                // to the left before its 1-day morph. During a morph, the explicit handoff effect
-                // above is the sole owner of selected-date updates.
-                if (latestMorphContext != null) return@collect
-                val target = programmaticTargetPage
-                if (target != null) {
-                    if (page == target) programmaticTargetPage = null
-                    return@collect
-                }
-                val date = page.toDayDate()
-                if (date != latestSelectedDate) {
-                    onDateSelected(date)
-                }
-            }
-    }
-
-    // Tapping a day header morphs between 3-day and 1-day. We capture which day was tapped and
-    // its column slot so the overlay can expand it FROM that position (not force it leftmost).
-    val onHeaderTap: (LocalDate) -> Unit = { day ->
-        if (isDayMode) {
-            // Collapse to multi-day: this day becomes the leftmost column and the next days slide in.
-            morphContext = DayMorphContext(
-                days = (0 until multiDayCount).map { day.plusDays(it.toLong()) },
-                expandSlot = 0,
-                targetMode = CalendarViewMode.ThreeDay,
-            )
-            scope.launch {
-                // First paint the affine overlay at the untouched 1-day geometry. Starting the
-                // width animation in the same frame as the tap lets pager/view state changes win
-                // the first draw and creates a visible snap before the actual morph.
-                withFrameNanos { }
-                timelineMode = CalendarViewMode.ThreeDay
-                onViewSelected(CalendarViewMode.ThreeDay)
-            }
-        } else {
-            // Expand this day to the 1-day view, zooming out of its current column in place.
-            val leftmost = visibleStartDate
-            val slot = ChronoUnit.DAYS.between(leftmost, day).toInt().coerceIn(0, multiDayCount - 1)
-            morphContext = DayMorphContext(
-                days = (0 until multiDayCount).map { leftmost.plusDays(it.toLong()) },
-                expandSlot = slot,
-                targetMode = CalendarViewMode.Day,
-            )
-            scope.launch {
-                // Keep one frame of the original 3-day positions before beginning the local
-                // animation. The persisted view setting follows the local mode instead of driving
-                // the animation, so a tapped middle/right date cannot be focused left first.
-                withFrameNanos { }
-                timelineMode = CalendarViewMode.Day
-                onViewSelected(CalendarViewMode.Day)
-            }
-        }
-    }
-
-    val todayDate = calendarTime.today
-    val todayPage = todayDate.toDayPage()
-    val currentLineGeometry = morphContext?.let { ctx ->
-        val todaySlot = ctx.days.indexOf(todayDate)
-        if (todaySlot >= 0 && dayGeometryViewportWidthPx > 0f) {
-            val columnWidthMulti = ((dayGeometryViewportWidthPx - (multiDayCount - 1f) * daySpacingPx) / multiDayCount).coerceAtLeast(1f)
-            val step = columnWidthMulti + daySpacingPx
-            val zoom = 1f + (dayGeometryViewportWidthPx / columnWidthMulti - 1f) * morphProgress
-            val e = ctx.expandSlot
-            CurrentLineGeometry(
-                leftPx = (todaySlot - e) * step * zoom + e * step * (1f - morphProgress),
-                widthPx = columnWidthMulti * zoom,
-            )
-        } else {
-            null
-        }
-    } ?: if (morphContext == null && dayViewportWidthPx > 0) {
-        CurrentLineGeometry(
-            leftPx = ((todayPage - pagerState.currentPage) - pagerState.currentPageOffsetFraction) * (dayWidthPx + daySpacingPx),
-            widthPx = dayWidthPx,
-        )
-    } else {
-        null
-    }
-    val currentLineVisible = currentLineGeometry?.let { geometry ->
-        geometry.leftPx < dayViewportWidthPx && geometry.leftPx + geometry.widthPx > 0f
-    } == true
-
-    Box(
-        Modifier
-            .fillMaxSize()
-            // The WHOLE 1-day timeline — time bar, all-day band and grid together — is the shared
-            // element for the month-cell <-> day morph, so the entire view scales up out of (and
-            // back into) the tapped month cell as one unit. Only enabled in 1-day mode: 3-day uses
-            // the affine overlay instead, and when this is disabled the modifier is a plain no-op.
-            // No solid backing: the real elements (grid, time bar, cards) fade in from transparency
-            // and the month cross-fades out underneath, so nothing flashes to a flat colour.
-            .morphBounds("dayblock-$monthMorphDay", enabled = isDayMode)
-            .clipToBounds(),
-    ) {
-        Row(Modifier.matchParentSize()) {
-            Column(Modifier.width(TimeSidebarWidth)) {
-                Box(
-                    modifier = Modifier
-                        .height(DayHeaderHeight)
-                        .fillMaxWidth(),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    androidx.compose.animation.AnimatedVisibility(
-                        visible = !isDayMode && state.multiDaySidebarControlsEnabled,
-                        enter = slideInHorizontally(
-                            initialOffsetX = { -it },
-                            animationSpec = tween(280, easing = MotionEmphasized),
-                        ) + fadeIn(animationSpec = tween(180, easing = MotionStandard)),
-                        exit = slideOutHorizontally(
-                            targetOffsetX = { -it },
-                            animationSpec = tween(280, easing = MotionEmphasized),
-                        ) + fadeOut(animationSpec = tween(180, easing = MotionStandard)),
-                        modifier = Modifier.fillMaxSize(),
-                    ) {
-                        MultiDayCountControls(
-                            dayCount = multiDayCount,
-                            onDayCountChanged = onMultiDayCountChanged,
-                        )
-                    }
-                }
-                Spacer(Modifier.height(animatedAllDayHeight))
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .weight(1f)
-                        .verticalScroll(timeScroll),
-                ) {
-                    Column {
-                        TimeSidebar(
-                            hours = DayStartHour..DayEndHour,
-                            hourHeightDp = hourHeightDp,
-                            onZoom = timelinePinchZoom,
-                        )
-                        if (timelineBottomInset > 0.dp) {
-                            Spacer(Modifier.height(timelineBottomInset + 20.dp))
-                        }
-                    }
-                }
-            }
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .onSizeChanged { dayViewportWidthPx = it.width }
-                    .verticalPinchZoom(
-                        topOffsetPx = with(density) { (DayHeaderHeight + animatedAllDayHeight).toPx() },
-                        viewportHeightPx = gridViewportHeightPx,
-                        onZoom = timelinePinchZoom,
-                    )
-                    // While the affine morph plays, clip to this area so the columns sliding off the
-                    // left edge disappear BEHIND the time bar instead of drawing over it. (Left
-                    // un-clipped in steady state so a card can still be dragged across day edges.)
-                    .then(if (morphContext != null) Modifier.clipToBounds() else Modifier),
-            ) {
-            var draggingDay by remember { mutableStateOf<LocalDate?>(null) }
-            // Enable per-item (event/task) morphs out of the month cell only for the real pager in
-            // 1-day mode and only when NOT doing the 3-day affine morph — so the selected day's
-            // cards pair up with the month pills, without the affine overlay's copies registering
-            // duplicate keys.
-            val itemMorphDay = if (morphContext == null && isDayMode) state.selectedDate else null
-            CompositionLocalProvider(
-                LocalMorphItemDay provides itemMorphDay,
-                LocalTimedDragReporter provides { date, active ->
-                    if (active) draggingDay = date
-                    else if (draggingDay == date) draggingDay = null
-                },
-            ) {
-            if (morphContext == null) {
-                TimedGridViewportLayer(
-                    visiblePages = pagerState.layoutInfo.visiblePagesInfo,
-                    pageWidthPx = dayWidthPx,
-                    hourHeightDp = hourHeightDp,
-                    topOffset = DayHeaderHeight + animatedAllDayHeight,
-                    scrollOffsetPx = timeScroll.value,
-                    modifier = Modifier
-                        .matchParentSize()
-                        .zIndex(0f),
-                )
-            }
-            HorizontalPager(
-                state = pagerState,
-                pageSize = PageSize.Fixed(dayWidthDp),
-                pageSpacing = DayColumnSpacing,
-                flingBehavior = PagerDefaults.flingBehavior(
-                    state = pagerState,
-                    pagerSnapDistance = PagerSnapDistance.atMost(multiDayCount.coerceAtLeast(3)),
-                ),
-                // Keep both neighbours composed so they're ready to slide in when the 1-day
-                // view widens out to 3-day, instead of popping in once they reach the viewport.
-                beyondViewportPageCount = multiDayCount.coerceAtLeast(2),
-                // Hidden (but kept composed, so scroll/zoom state survive) while a header-tap
-                // morph plays — the affine overlay below renders that transition instead.
-                modifier = Modifier
-                    .fillMaxSize()
-                    .zIndex(1f)
-                    .alpha(if (morphContext != null) 0f else 1f),
-            ) { page ->
-                val day = page.toDayDate()
-                Box(modifier = Modifier.zIndex(if (day == draggingDay) 50f else 0f)) {
-                DayPagerColumn(
-                    day = day,
-                    visibleStartDate = visibleStartDate,
-                    visibleEndDate = visibleEndDate,
-                    allDayHeight = animatedAllDayHeight,
-                    hourHeightDp = hourHeightDp,
-                    dayWidthPx = dayWidthPx,
-                    taskColorMode = state.taskColorMode,
-                    events = eventsByDay[day].orEmpty(),
-                    tasks = tasksByDay[day].orEmpty(),
-                    draftEvent = draftEvent?.takeIf { it.date == day },
-                    onDraftEventChanged = onDraftEventChanged,
-                    timeScroll = timeScroll,
-                    bottomObscuredHeight = timelineBottomInset,
-                    bodyScrollableState = bodyScrollableState,
-                    onDateSelected = onDateSelected,
-                    onDayHeaderClick = onHeaderTap,
-                    onSlotSelected = onSlotSelected,
-                    onDraftInteraction = onDraftInteraction,
-                    onDraftTap = onDraftTap,
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onEventMoved = onEventMoved,
-                    onTaskMoved = onTaskMoved,
-                    onEventMovedAllDay = onEventMovedAllDay,
-                    onTaskMovedAllDay = onTaskMovedAllDay,
-                    onDetail = onDetail,
-                    onGridViewportHeight = { gridViewportHeightPx = it },
-                    drawTimedGrid = false,
-                )
-                }
-            }
-            }
-            // In-place "zoom around the tapped column" overlay. While a header-tap morph plays it
-            // replaces the hidden pager so the tapped day expands FROM its own position (the pager
-            // can only ever anchor the leftmost page) and the other two columns slide off WITH
-            // their events. At progress 0 it matches the 3-day pager exactly and at 1 the 1-day
-            // pager, so swapping back to the real pager at either end is seamless. The expanding
-            // column's rect is mapped to the full viewport; every column scales around it.
-            morphContext?.let { ctx ->
-                val columnWidthMulti = ((dayGeometryViewportWidthPx - (multiDayCount - 1f) * daySpacingPx) / multiDayCount).coerceAtLeast(1f)
-                val step = columnWidthMulti + daySpacingPx
-                val zoom = 1f + (dayGeometryViewportWidthPx / columnWidthMulti - 1f) * morphProgress
-                val e = ctx.expandSlot
-                ctx.days.forEachIndexed { slot, day ->
-                    val leftPx = (slot - e) * step * zoom + e * step * (1f - morphProgress)
-                    val columnWidthPx = columnWidthMulti * zoom
-                    Box(
-                        modifier = Modifier
-                            .offset { IntOffset(leftPx.roundToInt(), 0) }
-                            .width(with(density) { columnWidthPx.toDp() })
-                            .fillMaxHeight()
-                            .zIndex(if (slot == e) 1f else 0f),
-                    ) {
-                        DayPagerColumn(
-                            day = day,
-                            visibleStartDate = day,
-                            visibleEndDate = day,
-                            allDayHeight = animatedAllDayHeight,
-                            hourHeightDp = hourHeightDp,
-                            dayWidthPx = columnWidthPx,
-                            taskColorMode = state.taskColorMode,
-                            events = eventsByDay[day].orEmpty(),
-                            tasks = tasksByDay[day].orEmpty(),
-                            draftEvent = draftEvent?.takeIf { it.date == day },
-                            onDraftEventChanged = onDraftEventChanged,
-                            timeScroll = timeScroll,
-                            bottomObscuredHeight = timelineBottomInset,
-                            bodyScrollableState = bodyScrollableState,
-                            onDateSelected = onDateSelected,
-                            onDayHeaderClick = onHeaderTap,
-                            onSlotSelected = onSlotSelected,
-                            onDraftInteraction = onDraftInteraction,
-                            onDraftTap = onDraftTap,
-                            onTaskStatusChanged = onTaskStatusChanged,
-                            onEventMoved = onEventMoved,
-                            onTaskMoved = onTaskMoved,
-                            onEventMovedAllDay = onEventMovedAllDay,
-                            onTaskMovedAllDay = onTaskMovedAllDay,
-                            onDetail = onDetail,
-                            onGridViewportHeight = {},
-                        )
-                    }
-                }
-                // The all-day band morphs WITH the columns: drive the same overlay off the affine
-                // geometry — anchor at the first column's animated left edge, with the animated
-                // gap between adjacent days as the step. Clipped to the day area by the parent, so
-                // chips on days sliding off the left vanish behind the time bar like the columns.
-                AllDayViewportOverlay(
-                    events = state.events,
-                    tasks = calendarTasks,
-                    taskColorMode = state.taskColorMode,
-                    anchorPage = ctx.days.first().toDayPage(),
-                    anchorOffsetPx = -e * step * zoom + e * step * (1f - morphProgress),
-                    dayWidthPx = columnWidthMulti * zoom,
-                    dayStepPx = step * zoom,
-                    viewportWidthPx = dayViewportWidthPx.toFloat(),
-                    topOffset = DayHeaderHeight,
-                    height = animatedAllDayHeight,
-                    hourHeightDp = hourHeightDp,
-                    timeScrollPx = timeScroll.value,
-                    defaultEventDurationMinutes = state.defaultEventDurationMinutes,
-                    draftEvent = draftEvent,
-                    onDraftTap = onDraftTap,
-                    onAllDaySlotSelected = onAllDaySlotSelected,
-                    onEventMoved = onEventMoved,
-                    onTaskMoved = onTaskMoved,
-                    onEventMovedAllDay = onEventMovedAllDay,
-                    onTaskMovedAllDay = onTaskMovedAllDay,
-                    maxVisibleItems = state.maxVisibleAllDayItems,
-                    expanded = allDayExpanded,
-                    onExpandedChange = { allDayExpanded = it },
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onDetail = onDetail,
-                    priorityPageCount = ctx.days.size,
-                )
-            }
-            val firstVisiblePageInfo = pagerState.layoutInfo.visiblePagesInfo.minByOrNull { it.offset }
-            // Normal pager-driven current-time line + all-day band, only when NOT morphing (during
-            // the morph the affine all-day overlay above takes over and the time line is hidden).
-            if (morphContext == null) {
-                AllDayViewportOverlay(
-                    events = state.events,
-                    tasks = calendarTasks,
-                    taskColorMode = state.taskColorMode,
-                    anchorPage = firstVisiblePageInfo?.index ?: pagerState.currentPage,
-                    anchorOffsetPx = firstVisiblePageInfo?.offset?.toFloat() ?: 0f,
-                    dayWidthPx = dayWidthPx,
-                    dayStepPx = dayWidthPx + daySpacingPx,
-                    viewportWidthPx = dayViewportWidthPx.toFloat(),
-                    topOffset = DayHeaderHeight,
-                    height = animatedAllDayHeight,
-                    hourHeightDp = hourHeightDp,
-                    timeScrollPx = timeScroll.value,
-                    defaultEventDurationMinutes = state.defaultEventDurationMinutes,
-                    draftEvent = draftEvent,
-                    onDraftTap = onDraftTap,
-                    onAllDaySlotSelected = onAllDaySlotSelected,
-                    onEventMoved = onEventMoved,
-                    onTaskMoved = onTaskMoved,
-                    onEventMovedAllDay = onEventMovedAllDay,
-                    onTaskMovedAllDay = onTaskMovedAllDay,
-                    maxVisibleItems = state.maxVisibleAllDayItems,
-                    expanded = allDayExpanded,
-                    onExpandedChange = { allDayExpanded = it },
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onDetail = onDetail,
-                    priorityPageCount = targetDayCount,
-                )
-            }
-            }
-        }
-        if (allDayHasOverflow || allDayExpanded) {
-            Box(
-                modifier = Modifier
-                    .offset(y = DayHeaderHeight)
-                    .width(TimeSidebarWidth)
-                    .height(animatedAllDayHeight)
-                    .padding(bottom = 4.dp)
-                    .zIndex(14f)
-                    .clickable(
-                        interactionSource = remember { MutableInteractionSource() },
-                        indication = null,
-                        onClick = { allDayExpanded = !allDayExpanded },
-                    ),
-                contentAlignment = Alignment.BottomCenter,
-            ) {
-                Icon(
-                    imageVector = Icons.Default.KeyboardArrowDown,
-                    contentDescription = if (allDayExpanded) stringResource(R.string.collapse_all_day_items) else stringResource(R.string.expand_all_day_items),
-                    tint = WarmInk.copy(alpha = 0.82f),
-                    modifier = Modifier
-                        .size(28.dp)
-                        .graphicsLayer {
-                            rotationZ = allDayArrowRotation
-                            val arrowScale = 0.94f + 0.06f * abs(allDayArrowRotation - 90f) / 90f
-                            scaleX = arrowScale
-                            scaleY = arrowScale
-                        },
-                )
-            }
-        }
-        if (currentLineVisible && currentLineGeometry != null) {
-            CurrentTimeLine(
-                now = now,
-                todayPageOffset = with(density) { currentLineGeometry.leftPx.toDp() },
-                dayWidth = with(density) { currentLineGeometry.widthPx.toDp() },
-                hourHeightDp = hourHeightDp,
-                topOffset = DayHeaderHeight + animatedAllDayHeight,
-                scrollOffset = with(density) { timeScroll.value.toDp() },
-                modifier = Modifier
-                    .matchParentSize()
-                    .zIndex(30f),
-            )
-        }
-    }
-}
-
-@Composable
-private fun MultiDayCountControls(
-    dayCount: Int,
-    onDayCountChanged: (Int) -> Unit,
-) {
-    val safeDayCount = dayCount.coerceMultiDayCount()
-    val canIncrease = safeDayCount < MAX_MULTI_DAY_COUNT
-    val canDecrease = safeDayCount > MIN_MULTI_DAY_COUNT
-    val railShape = RoundedCornerShape(topStart = 0.dp, bottomStart = 0.dp, topEnd = 16.dp, bottomEnd = 16.dp)
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .clip(railShape)
-            .background(multiDayCountRailColor())
-            .fillMaxWidth()
-            .padding(end = 2.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.spacedBy(8.dp, Alignment.CenterVertically),
-    ) {
-        MultiDayCountIconButton(
-            icon = Icons.Default.Add,
-            contentDescription = stringResource(R.string.increase_multi_day_count),
-            enabled = canIncrease,
-            onClick = { onDayCountChanged(safeDayCount + 1) },
-        )
-        MultiDayCountIconButton(
-            icon = Icons.Default.Remove,
-            contentDescription = stringResource(R.string.decrease_multi_day_count),
-            enabled = canDecrease,
-            onClick = { onDayCountChanged(safeDayCount - 1) },
-        )
-    }
-}
-
-@Composable
-private fun MultiDayCountIconButton(
-    icon: ImageVector,
-    contentDescription: String,
-    enabled: Boolean,
-    onClick: () -> Unit,
-) {
-    val tint = if (enabled) WarmInk else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.28f)
-    Box(
-        modifier = Modifier
-            .size(24.dp)
-            .clip(CircleShape)
-            .clickable(
-                enabled = enabled,
-                interactionSource = remember { MutableInteractionSource() },
-                indication = null,
-                onClick = onClick,
-            ),
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            icon,
-            contentDescription = contentDescription,
-            tint = tint,
-            modifier = Modifier.size(18.dp),
-        )
-    }
-}
-
-@Composable
-private fun multiDayCountRailColor(): Color =
-    if (MaterialTheme.colorScheme.background.isDark()) {
-        MaterialTheme.colorScheme.surface.blendWith(Color.White, 0.08f)
-    } else {
-        Color.White
-    }
-
-/** A header-tap morph in progress: which multi-day columns are involved and which one
- * expands to fill / collapses from the 1-day view. */
-private data class DayMorphContext(
-    val days: List<LocalDate>,
-    val expandSlot: Int,
-    val targetMode: CalendarViewMode,
-)
-
-private data class CurrentLineGeometry(
-    val leftPx: Float,
-    val widthPx: Float,
-)
-
-@Composable
-private fun TimedGridViewportLayer(
-    visiblePages: List<androidx.compose.foundation.pager.PageInfo>,
-    pageWidthPx: Float,
-    hourHeightDp: Float,
-    topOffset: Dp,
-    scrollOffsetPx: Int,
-    modifier: Modifier = Modifier,
-) {
-    val density = LocalDensity.current
-    val gridColor = WarmGrid
-    Canvas(modifier) {
-        val topOffsetPx = with(density) { topOffset.toPx() }
-        val rowHeight = with(density) { hourHeightDp.dp.toPx() }
-        val cellGap = HourCellGap.toPx()
-        val cellHeight = (rowHeight - cellGap).coerceAtLeast(0f)
-        val cellInset = cellGap / 2f
-        val radius = 10.dp.toPx()
-        clipRect(left = 0f, top = topOffsetPx, right = size.width, bottom = size.height) {
-            visiblePages.forEach { page ->
-                val left = page.offset.toFloat()
-                repeat(DayEndHour - DayStartHour + 1) { index ->
-                    val top = topOffsetPx - scrollOffsetPx + index * rowHeight + cellInset
-                    if (top > size.height || top + cellHeight < topOffsetPx) return@repeat
-                    drawRoundRect(
-                        color = gridColor,
-                        topLeft = Offset(left, top),
-                        size = Size(width = pageWidthPx, height = cellHeight),
-                        cornerRadius = CornerRadius(radius, radius),
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DayHeader(day: LocalDate, selected: Boolean, modifier: Modifier = Modifier) {
-    BoxWithConstraints(modifier = modifier, contentAlignment = Alignment.Center) {
-        val compact = maxWidth < 44.dp
-        val circleSize = if (compact) 32.dp else 40.dp
-        val weekdayFont = if (compact) 11.sp else 14.sp
-        val weekdayLine = if (compact) 13.sp else 16.sp
-        val dayFont = if (compact) 17.sp else 22.sp
-        val dayLine = if (compact) 19.sp else 24.sp
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            Text(
-                day.format(DateTimeFormatter.ofPattern("EE", LocalAppLocale.current)).replace(".", ""),
-                color = if (selected) WarmBrown else WarmInk,
-                fontWeight = FontWeight.Bold,
-                fontSize = weekdayFont,
-                lineHeight = weekdayLine,
-                maxLines = 1,
-                overflow = TextOverflow.Clip,
-            )
-            Box(
-                modifier = Modifier
-                    .size(circleSize)
-                    .clip(CircleShape)
-                    .background(if (selected) WarmBrown else Color.Transparent),
-                contentAlignment = Alignment.Center,
-            ) {
-                Text(
-                    day.dayOfMonth.toString(),
-                    color = if (selected) Color.White else WarmInk,
-                    fontSize = dayFont,
-                    lineHeight = dayLine,
-                    fontWeight = FontWeight.SemiBold,
-                    maxLines = 1,
-                )
-            }
-        }
-    }
-}
-
-/**
- * Reports when a timed card inside a 3-day pager page starts/stops being dragged.
- * The pager parent uses this to boost the dragging page's zIndex so the card can
- * visually escape the page bounds — without this, dragging an event to a future
- * day renders behind the next page's background, since pager pages are drawn in
- * index order by default.
- */
-private val LocalTimedDragReporter = compositionLocalOf<(LocalDate, Boolean) -> Unit> { { _, _ -> } }
 internal val LocalSheetHeaderDragModifier = compositionLocalOf<Modifier> { Modifier }
 private val LocalPendingMutations = compositionLocalOf<List<PendingMutationEntity>> { emptyList() }
 
@@ -6280,7 +5453,7 @@ private val LocalMorphAnimatedVisibilityScope = compositionLocalOf<AnimatedVisib
 // find everything it gates.)
 private const val MorphSlowFactor = 1
 
-private const val MorphDurationMs = 440 * MorphSlowFactor
+internal const val MorphDurationMs = 440 * MorphSlowFactor
 
 // Fades for the month-cell <-> day container transform (the only thing morphBounds now drives;
 // 3-day<->1-day is the affine overlay). The day block scales from/into the cell via ScaleToBounds;
@@ -6294,7 +5467,7 @@ private const val MorphDurationMs = 440 * MorphSlowFactor
 // The day whose event/task cards should register per-item shared elements (so they morph out of
 // the tapped month cell's pills). Set only while a month<->1-day morph is applicable; null in the
 // 3-day affine overlay and otherwise, so cards don't register duplicate/ stray shared elements.
-private val LocalMorphItemDay = compositionLocalOf<LocalDate?> { null }
+internal val LocalMorphItemDay = compositionLocalOf<LocalDate?> { null }
 
 // Option B (per-event pill<->card morph). OFF by deliberate choice. A per-item element travels its
 // OWN bounds path (pill -> card) whose size delta is far smaller than the grid's (a pill fills a big
@@ -6307,7 +5480,7 @@ private val LocalMorphItemDay = compositionLocalOf<LocalDate?> { null }
 // the month pills cross-fade into the full cards as the whole day zooms — the real Material/Google
 // Calendar container transform. The per-item plumbing below stays behind this flag in case we ever
 // want to revisit the independent morph; flip to true to bring it back (with the geometric caveat).
-private const val EnablePerItemMorph = false
+internal const val EnablePerItemMorph = false
 
 private val MorphContentFadeMs = MorphDurationMs * 3 / 10
 private val MorphDayEnter: EnterTransition = fadeIn(tween(MorphContentFadeMs, easing = MorphEasing))
@@ -6325,8 +5498,8 @@ private val MorphCellExit: ExitTransition = fadeOut(tween(MorphContentFadeMs, ea
 // with it) is visible the entire time, matching the smooth scale-DOWN. The PILL side is the
 // opposite (only present at the cell-sized end), so it never lingers as a ballooning copy.
 private val MorphItemFadeMs = MorphDurationMs / 6
-private val MorphItemCardEnter: EnterTransition = fadeIn(tween(MorphItemFadeMs, easing = MorphEasing))
-private val MorphItemCardExit: ExitTransition =
+internal val MorphItemCardEnter: EnterTransition = fadeIn(tween(MorphItemFadeMs, easing = MorphEasing))
+internal val MorphItemCardExit: ExitTransition =
     fadeOut(tween(MorphItemFadeMs, delayMillis = MorphDurationMs - MorphItemFadeMs, easing = MorphEasing))
 private val MorphItemPillEnter: EnterTransition =
     fadeIn(tween(MorphItemFadeMs, delayMillis = MorphDurationMs - MorphItemFadeMs, easing = MorphEasing))
@@ -6351,7 +5524,7 @@ private val MorphItemPillExit: ExitTransition = fadeOut(tween(MorphItemFadeMs, e
 // strongly held back while short all-day chips (which barely diverge) are left almost untouched.
 //
 // The shrinking direction keeps the plain MorphEasing so the already-perfect reverse is unchanged.
-private val MorphItemBoundsTransform = androidx.compose.animation.BoundsTransform { initial, target ->
+internal val MorphItemBoundsTransform = androidx.compose.animation.BoundsTransform { initial, target ->
     if (target.height >= initial.height) {
         val ratio = (initial.height / target.height).coerceIn(0f, 0.92f)
         val gridTracking = androidx.compose.animation.core.Easing { f ->
@@ -6417,7 +5590,7 @@ private fun CalendarViewMode.group(): CalendarViewGroup = when (this) {
  * fade with their scene). No-op entirely when not inside the morph container.
  */
 @Composable
-private fun Modifier.morphBounds(
+internal fun Modifier.morphBounds(
     key: Any,
     enabled: Boolean = true,
     enter: EnterTransition = MorphDayEnter,
@@ -6560,2237 +5733,10 @@ private val MorphRoundedClip: SharedTransitionScope.OverlayClip =
         }
     }
 
-@Composable
-private fun DayPagerColumn(
-    day: LocalDate,
-    visibleStartDate: LocalDate,
-    visibleEndDate: LocalDate,
-    allDayHeight: Dp,
-    events: List<EventEntity>,
-    tasks: List<TaskEntity>,
-    draftEvent: DraftEventSelection?,
-    onDraftEventChanged: (DraftEventSelection) -> Unit,
-    onDraftInteraction: () -> Unit,
-    onDraftTap: () -> Unit,
-    hourHeightDp: Float,
-    dayWidthPx: Float,
-    timeScroll: androidx.compose.foundation.ScrollState,
-    bottomObscuredHeight: Dp,
-    bodyScrollableState: androidx.compose.foundation.gestures.ScrollableState,
-    onDateSelected: (LocalDate) -> Unit,
-    onDayHeaderClick: (LocalDate) -> Unit,
-    onSlotSelected: (LocalDate, LocalTime) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEventMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onTaskMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onEventMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onTaskMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onDetail: (DetailSheet) -> Unit,
-    onGridViewportHeight: (Int) -> Unit,
-    taskColorMode: TaskColorMode,
-    drawTimedGrid: Boolean = true,
-) {
-    // Just the day's own content (header + timed grid + cards). The month-cell <-> day morph is
-    // now applied one level up, on the whole timeline Row (so the time bar and all-day band scale
-    // out of the tapped cell too); the 3-day <-> 1-day morph is the affine overlay. So this column
-    // no longer carries a shared element itself.
-    val headerInteraction = remember(day) { MutableInteractionSource() }
-    val density = LocalDensity.current
-    var localGridViewportHeightPx by remember { mutableStateOf(0) }
-    val bottomObscuredPx = with(density) { bottomObscuredHeight.roundToPx() }
-    val effectiveGridViewportHeightPx = (localGridViewportHeightPx - bottomObscuredPx).coerceAtLeast(0)
-    LaunchedEffect(effectiveGridViewportHeightPx) {
-        if (effectiveGridViewportHeightPx > 0) onGridViewportHeight(effectiveGridViewportHeightPx)
-    }
-    Column(Modifier.fillMaxSize()) {
-        DayHeader(
-            day = day,
-            selected = day == LocalCalendarTimeSnapshot.current.today,
-            modifier = Modifier
-                .zIndex(4f)
-                .height(DayHeaderHeight)
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(22.dp))
-                .clickable(
-                    interactionSource = headerInteraction,
-                    indication = null,
-                    onClick = { onDayHeaderClick(day) },
-                ),
-        )
-        Spacer(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(allDayHeight)
-                .zIndex(4f),
-        )
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f)
-                .onSizeChanged {
-                    localGridViewportHeightPx = it.height
-                    onGridViewportHeight((it.height - bottomObscuredPx).coerceAtLeast(0))
-                }
-                .verticalClipAllowHorizontalOverflow()
-                .scrollable(state = bodyScrollableState, orientation = Orientation.Vertical),
-        ) {
-            Box(
-                modifier = Modifier
-                    .matchParentSize()
-                    .pointerInput(day, hourHeightDp, timeScroll.value) {
-                        val rowHeightPx = with(density) { hourHeightDp.dp.toPx() }
-                        detectTapGestures { offset ->
-                            val scrolledY = offset.y + timeScroll.value
-                            val totalVisibleMinutes = ((scrolledY / rowHeightPx) * 60f)
-                                .roundToInt()
-                                .snapDraftMinute()
-                            val minuteOfDay = (DayStartHour * 60 + totalVisibleMinutes)
-                                .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - DraftMinDurationMinutes)
-                            onSlotSelected(day, minuteOfDay.toDraftLocalTime())
-                        }
-                    },
-            )
-            DayTimedColumn(
-                day = day,
-                hourHeightDp = hourHeightDp,
-                dayWidthPx = dayWidthPx,
-                taskColorMode = taskColorMode,
-                events = events,
-                tasks = tasks,
-                draftEvent = draftEvent,
-                onDraftEventChanged = onDraftEventChanged,
-                onDraftInteraction = onDraftInteraction,
-                onDraftTap = onDraftTap,
-                timeScroll = timeScroll,
-                gridViewportHeightPx = effectiveGridViewportHeightPx,
-                onSlotSelected = onSlotSelected,
-                onTaskStatusChanged = onTaskStatusChanged,
-                onEventMoved = onEventMoved,
-                onTaskMoved = onTaskMoved,
-                onEventMovedAllDay = onEventMovedAllDay,
-                onTaskMovedAllDay = onTaskMovedAllDay,
-                onDetail = onDetail,
-                drawGrid = drawTimedGrid,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .offset { IntOffset(0, -timeScroll.value) }
-                    .height(((DayEndHour - DayStartHour + 1) * hourHeightDp).dp),
-            )
-        }
-    }
-}
 
-@Composable
-private fun AllDayViewportOverlay(
-    events: List<EventEntity>,
-    tasks: List<TaskEntity>,
-    taskColorMode: TaskColorMode,
-    anchorPage: Int,
-    anchorOffsetPx: Float,
-    dayWidthPx: Float,
-    dayStepPx: Float,
-    viewportWidthPx: Float,
-    topOffset: Dp,
-    height: Dp,
-    hourHeightDp: Float,
-    timeScrollPx: Int,
-    defaultEventDurationMinutes: Int,
-    draftEvent: DraftEventSelection?,
-    onDraftTap: () -> Unit,
-    onAllDaySlotSelected: (LocalDate) -> Unit,
-    onEventMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onTaskMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onEventMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onTaskMovedAllDay: (String, Long, LocalDate) -> Unit,
-    maxVisibleItems: Int,
-    expanded: Boolean,
-    onExpandedChange: (Boolean) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onDetail: (DetailSheet) -> Unit,
-    priorityPageCount: Int,
-) {
-    if (height <= 0.dp || viewportWidthPx <= 0f || dayStepPx <= 0f) return
-    val bufferStartPage = (anchorPage - 10).coerceAtLeast(0)
-    val bufferEndPage = (anchorPage + 14).coerceAtMost(DayPagerPageCount - 1)
-    val daySpacingPx = (dayStepPx - dayWidthPx).coerceAtLeast(0f)
-    val viewportPages = remember(anchorPage, anchorOffsetPx, dayWidthPx, dayStepPx, viewportWidthPx, bufferStartPage, bufferEndPage) {
-        (bufferStartPage..bufferEndPage).filter { page ->
-            val left = anchorOffsetPx + (page - anchorPage) * dayStepPx
-            left < viewportWidthPx + daySpacingPx && left + dayWidthPx + daySpacingPx > 0f
-        }
-    }
-    val visibleStartPage = viewportPages.minOrNull() ?: anchorPage.coerceIn(bufferStartPage, bufferEndPage)
-    val visibleEndPage = viewportPages.maxOrNull() ?: visibleStartPage
-    val corePageCount = priorityPageCount.coerceIn(1, MAX_MULTI_DAY_COUNT)
-    val priorityPages = remember(anchorPage, anchorOffsetPx, dayWidthPx, dayStepPx, viewportWidthPx, visibleStartPage, visibleEndPage, corePageCount) {
-        val visiblePages = (visibleStartPage..visibleEndPage).toList()
-        if (visiblePages.size <= corePageCount) {
-            visiblePages
-        } else {
-            val coverageByPage = visiblePages.associateWith { page ->
-                val left = anchorOffsetPx + (page - anchorPage) * dayStepPx
-                val right = left + dayWidthPx
-                (min(right, viewportWidthPx) - max(left, 0f)).coerceIn(0f, dayWidthPx)
-            }
-            val minStart = visiblePages.first()
-            val maxStart = visiblePages.last() - corePageCount + 1
-            (minStart..maxStart)
-                .maxWithOrNull(
-                    compareBy<Int> { start ->
-                        (start until start + corePageCount).sumOf { page -> (coverageByPage[page] ?: 0f).toDouble() }
-                    }.thenBy { start ->
-                        -abs(start - anchorPage)
-                    },
-                )
-                ?.let { start -> (start until start + corePageCount).toList() }
-                ?: visiblePages.take(corePageCount)
-        }
-    }
-    val priorityStartPage = priorityPages.firstOrNull() ?: visibleStartPage
-    val priorityEndPage = priorityPages.lastOrNull() ?: visibleEndPage
-    val layoutStartPage = visibleStartPage
-    val layoutEndPage = visibleEndPage
-    val overlayItems = remember(events, tasks, layoutStartPage, layoutEndPage, priorityStartPage, priorityEndPage, taskColorMode) {
-        buildAllDayOverlayItems(
-            events = events,
-            tasks = tasks,
-            taskColorMode = taskColorMode,
-            visibleStartPage = layoutStartPage,
-            visibleEndPage = layoutEndPage,
-            priorityStartPage = priorityStartPage,
-            priorityEndPage = priorityEndPage,
-        )
-    }
-    val rowCount = overlayItems.maxOfOrNull { it.lane + 1 } ?: 0
-    val pageItemsByPage = remember(overlayItems, layoutStartPage, layoutEndPage, priorityStartPage, priorityEndPage) {
-        (layoutStartPage..layoutEndPage).associateWith { page ->
-            overlayItems
-                .filter { page in it.startPage..it.endPage }
-                .sortedWith(allDayCollapsedPageItemComparator(priorityStartPage, priorityEndPage))
-        }
-    }
-    val hasCollapsedOverflow = maxVisibleItems < rowCount &&
-        (visibleStartPage..visibleEndPage).any { pageItemsByPage[it].orEmpty().size > maxVisibleItems }
-    val collapsedVisibleItemLimit = when {
-        maxVisibleItems <= 0 && rowCount > 0 -> 0
-        hasCollapsedOverflow -> (maxVisibleItems - 1).coerceAtLeast(0)
-        else -> maxVisibleItems
-    }
-    val overflowLane = if (maxVisibleItems <= 0) 0 else collapsedVisibleItemLimit
-    val density = LocalDensity.current
-    val expansionProgress by animateFloatAsState(
-        targetValue = if (expanded) 1f else 0f,
-        animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-        label = "allDayOverflowExpansionProgress",
-    )
-    val renderExpandedItems = expanded || expansionProgress > 0.01f
-    val collapsedLayout = remember(overlayItems, pageItemsByPage, layoutStartPage, layoutEndPage, maxVisibleItems, collapsedVisibleItemLimit) {
-        buildCollapsedAllDayLayout(
-            overlayItems = overlayItems,
-            pageItemsByPage = pageItemsByPage,
-            visibleStartPage = layoutStartPage,
-            visibleEndPage = layoutEndPage,
-            maxVisibleItems = maxVisibleItems,
-            collapsedVisibleItemLimit = collapsedVisibleItemLimit,
-        )
-    }
-    val collapsedSegments = collapsedLayout.segments
-    val continuationSegments = collapsedLayout.continuations
-    val showCollapsedOverflowChips = hasCollapsedOverflow && !expanded && expansionProgress <= 0.08f
-    var draggingAllDayItemId by remember { mutableStateOf<String?>(null) }
-    fun moveAllDayItemToTimed(item: AllDayOverlayItem, date: LocalDate, start: LocalTime, end: LocalTime) {
-        item.event?.let { event ->
-            onEventMoved(event.resourceHref, event.startsAtMillis, date, start, end)
-        }
-        item.task?.let { task ->
-            onTaskMoved(task.resourceHref, task.startAtMillis ?: task.dueAtMillis ?: System.currentTimeMillis(), date, start, end)
-        }
-    }
-    fun moveAllDayItemToAllDay(item: AllDayOverlayItem, date: LocalDate) {
-        item.event?.let { event ->
-            onEventMovedAllDay(event.resourceHref, event.startsAtMillis, date)
-        }
-        item.task?.let { task ->
-            onTaskMovedAllDay(task.resourceHref, task.startAtMillis ?: task.dueAtMillis ?: System.currentTimeMillis(), date)
-        }
-    }
-    val hiddenPages = if (!showCollapsedOverflowChips) {
-        emptyMap()
-    } else {
-        (visibleStartPage..visibleEndPage).mapNotNull { page ->
-            val pageItems = pageItemsByPage[page].orEmpty()
-            if (pageItems.size <= maxVisibleItems) return@mapNotNull null
-            val hiddenItems = pageItems.drop(collapsedVisibleItemLimit)
-            if (hiddenItems.isEmpty()) null else page to hiddenItems
-        }.toMap()
-    }
-    Box(
-        modifier = Modifier
-            .offset(y = topOffset)
-            .fillMaxWidth()
-            .height(height)
-            .then(if (draggingAllDayItemId == null) Modifier.verticalClipAllowHorizontalOverflow() else Modifier)
-            .zIndex(9f),
-    ) {
-        Box(
-            modifier = Modifier
-                .matchParentSize()
-                .pointerInput(anchorPage, anchorOffsetPx, dayStepPx) {
-                    detectTapGestures { offset ->
-                        val pageDelta = floor((offset.x - anchorOffsetPx) / dayStepPx).toInt()
-                        val tappedPage = (anchorPage + pageDelta).coerceIn(0, DayPagerPageCount - 1)
-                        onAllDaySlotSelected(tappedPage.toDayDate())
-                    }
-                },
-        )
-        val expandedRenderItems = if (renderExpandedItems) overlayItems else emptyList()
-        expandedRenderItems.forEach { item ->
-            val startX = anchorOffsetPx + (item.startPage - anchorPage) * dayStepPx
-            val endX = anchorOffsetPx + (item.endPage - anchorPage) * dayStepPx + dayWidthPx
-            val visibleLeft = startX.coerceAtLeast(0f)
-            val visibleRight = endX.coerceAtMost(viewportWidthPx)
-            val visibleWidth = visibleRight - visibleLeft
-            if (visibleWidth <= 1f) return@forEach
-            key("expanded-all-day:${item.id}") {
-                val animatedBaseLane by animateFloatAsState(
-                    targetValue = item.lane.toFloat(),
-                    animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-                    label = "expandedAllDayLane",
-                )
-                val isOverflowItem = item.lane >= collapsedVisibleItemLimit
-                val animatedLane = if (isOverflowItem) {
-                    overflowLane + (animatedBaseLane - overflowLane) * expansionProgress
-                } else {
-                    animatedBaseLane
-                }
-                val itemAlpha = if (isOverflowItem) expansionProgress else 1f
-                val chipTopDp = 7.dp + (animatedLane * 29f).dp
-                AllDayViewportChip(
-                    item = item,
-                    chipLeftPx = visibleLeft,
-                    chipTopDp = chipTopDp,
-                    allDayHeight = height,
-                    hourHeightDp = hourHeightDp,
-                    timeScrollPx = timeScrollPx,
-                    anchorPage = anchorPage,
-                    anchorOffsetPx = anchorOffsetPx,
-                    dayStepPx = dayStepPx,
-                    defaultDurationMinutes = if (item.task != null) (DEFAULT_TASK_DURATION_MILLIS / 60_000L).toInt() else defaultEventDurationMinutes,
-                    modifier = Modifier
-                        .offset {
-                            IntOffset(
-                                x = visibleLeft.roundToInt(),
-                                y = with(density) { chipTopDp.roundToPx() },
-                            )
-                        }
-                        .graphicsLayer {
-                            alpha = itemAlpha
-                            scaleY = if (isOverflowItem) 0.94f + 0.06f * expansionProgress else 1f
-                        }
-                        .width(with(density) { visibleWidth.toDp() })
-                        .height(24.dp),
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onDetail = onDetail,
-                    onMoveToTimed = ::moveAllDayItemToTimed,
-                    onMoveToAllDay = ::moveAllDayItemToAllDay,
-                    onDragStateChanged = { active -> draggingAllDayItemId = if (active) item.id else null },
-                )
-            }
-        }
-        if (!renderExpandedItems) {
-            fun pageVisibility(page: Int): Float {
-                val pageLeft = anchorOffsetPx + (page - anchorPage) * dayStepPx
-                val pageRight = pageLeft + dayWidthPx
-                return ((min(pageRight, viewportWidthPx) - max(pageLeft, 0f)) / dayWidthPx)
-                    .coerceIn(0f, 1f)
-            }
 
-            continuationSegments.forEach { segment ->
-                val left = anchorOffsetPx + (segment.page - anchorPage) * dayStepPx
-                val spacing = (dayStepPx - dayWidthPx).coerceAtLeast(0f)
-                val overlap = with(density) { 9.dp.toPx() }
-                val edgeReach = dayWidthPx * 0.18f
-                val fromPreviousAlpha = if (segment.fromPrevious) pageVisibility(segment.page - 1) else 0f
-                val toNextAlpha = if (segment.toNext) pageVisibility(segment.page + 1) else 0f
-                val segmentLeft = when {
-                    segment.fromPrevious -> left - spacing - overlap
-                    segment.toNext -> left + dayWidthPx - edgeReach
-                    else -> left
-                }
-                val segmentRight = when {
-                    segment.toNext -> left + dayWidthPx + spacing + overlap
-                    segment.fromPrevious -> left + edgeReach
-                    else -> left + dayWidthPx
-                }
-                val visibleLeft = segmentLeft.coerceAtLeast(0f)
-                val visibleRight = segmentRight.coerceAtMost(viewportWidthPx)
-                val visibleWidth = visibleRight - visibleLeft
-                if (visibleWidth <= 1f) return@forEach
-                val physicalFadeWidth = (spacing + overlap + edgeReach).coerceAtLeast(1f)
-                val sideReachFraction = (physicalFadeWidth / visibleWidth).coerceIn(
-                    0.05f,
-                    if (segment.fromPrevious && segment.toNext) 0.48f else 1f,
-                )
-                key("all-day-continuation:${segment.item.id}:${segment.page}") {
-                    val animatedLane by animateFloatAsState(
-                        targetValue = segment.lane.toFloat(),
-                        animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-                        label = "allDayContinuationLane",
-                    )
-                    AllDayContinuationChip(
-                        segment = segment,
-                        fromPreviousAlpha = fromPreviousAlpha,
-                        toNextAlpha = toNextAlpha,
-                        sideReachFraction = sideReachFraction,
-                        modifier = Modifier
-                            .offset {
-                                IntOffset(
-                                    x = visibleLeft.roundToInt(),
-                                    y = with(density) { (7.dp + (animatedLane * 29f).dp).roundToPx() },
-                                )
-                            }
-                            .width(with(density) { visibleWidth.toDp() })
-                            .height(24.dp),
-                    )
-                }
-            }
-            collapsedSegments.forEach { segment ->
-                val item = segment.item
-                val startX = anchorOffsetPx + (segment.startPage - anchorPage) * dayStepPx
-                val endX = anchorOffsetPx + (segment.endPage - anchorPage) * dayStepPx + dayWidthPx
-                val visibleLeft = startX.coerceAtLeast(0f)
-                val visibleRight = endX.coerceAtMost(viewportWidthPx)
-                val visibleWidth = visibleRight - visibleLeft
-                if (visibleWidth <= 1f) return@forEach
-                val segmentOrdinal = collapsedSegments
-                    .asSequence()
-                    .filter { it.item.id == segment.item.id && it.startPage < segment.startPage }
-                    .count()
-                key("collapsed-all-day:${segment.item.id}:$segmentOrdinal") {
-                    val animatedLane by animateFloatAsState(
-                        targetValue = segment.lane.toFloat(),
-                        animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-                        label = "collapsedAllDayLane",
-                    )
-                    val chipTopDp = 7.dp + (animatedLane * 29f).dp
-                    AllDayViewportChip(
-                        item = item,
-                        chipLeftPx = visibleLeft,
-                        chipTopDp = chipTopDp,
-                        allDayHeight = height,
-                        hourHeightDp = hourHeightDp,
-                        timeScrollPx = timeScrollPx,
-                        anchorPage = anchorPage,
-                        anchorOffsetPx = anchorOffsetPx,
-                        dayStepPx = dayStepPx,
-                        defaultDurationMinutes = if (item.task != null) (DEFAULT_TASK_DURATION_MILLIS / 60_000L).toInt() else defaultEventDurationMinutes,
-                        modifier = Modifier
-                            .offset {
-                                IntOffset(
-                                    x = visibleLeft.roundToInt(),
-                                    y = with(density) { chipTopDp.roundToPx() },
-                                )
-                            }
-                            .width(with(density) { visibleWidth.toDp() })
-                            .height(24.dp),
-                        onTaskStatusChanged = onTaskStatusChanged,
-                        onDetail = onDetail,
-                        onMoveToTimed = ::moveAllDayItemToTimed,
-                        onMoveToAllDay = ::moveAllDayItemToAllDay,
-                        onDragStateChanged = { active -> draggingAllDayItemId = if (active) item.id else null },
-                    )
-                }
-            }
-        }
-        hiddenPages.forEach { (page, hiddenItems) ->
-            val left = anchorOffsetPx + (page - anchorPage) * dayStepPx
-            val visibleLeft = left.coerceAtLeast(0f)
-            val visibleRight = (left + dayWidthPx).coerceAtMost(viewportWidthPx)
-            if (visibleRight - visibleLeft <= 1f) return@forEach
-            key("all-day-overflow:$page") {
-                val animatedOverflowLane by animateFloatAsState(
-                    targetValue = overflowLane.toFloat(),
-                    animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-                    label = "allDayOverflowLane",
-                )
-                AllDayOverflowChip(
-                    modifier = Modifier
-                        .offset {
-                            IntOffset(
-                                x = visibleLeft.roundToInt(),
-                                y = with(density) { (7.dp + (animatedOverflowLane * 29f).dp).roundToPx() },
-                            )
-                        }
-                        .width(with(density) { (visibleRight - visibleLeft).toDp() })
-                        .height(22.dp),
-                    hiddenItems = hiddenItems,
-                    onClick = { onExpandedChange(true) },
-                )
-            }
-        }
-        draftEvent?.takeIf { it.allDay }?.let { draft ->
-            val draftPage = draft.date.toDayPage()
-            val left = anchorOffsetPx + (draftPage - anchorPage) * dayStepPx
-            val visibleLeft = left.coerceAtLeast(0f)
-            val visibleRight = (left + dayWidthPx).coerceAtMost(viewportWidthPx)
-            val draftLane = if (renderExpandedItems) {
-                firstFreeLaneIndex(
-                    overlayItems
-                        .filter { draftPage in it.startPage..it.endPage }
-                        .map { it.lane }
-                        .toSet(),
-                )
-            } else {
-                val pageItems = pageItemsByPage[draftPage].orEmpty()
-                val visibleDraftSlots = when {
-                    maxVisibleItems <= 0 -> 0
-                    pageItems.size > maxVisibleItems -> collapsedVisibleItemLimit
-                    else -> maxVisibleItems
-                }
-                if (pageItems.size < visibleDraftSlots) pageItems.size else null
-            }
-            if (visibleRight - visibleLeft > 1f && draftLane != null) {
-                val animatedDraftLane by animateFloatAsState(
-                    targetValue = draftLane.toFloat(),
-                    animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-                    label = "allDayDraftLane",
-                )
-                Box(
-                    modifier = Modifier
-                        .offset {
-                            IntOffset(
-                                x = visibleLeft.roundToInt(),
-                                y = with(density) { (7.dp + (animatedDraftLane * 29f).dp).roundToPx() },
-                            )
-                        }
-                        .width(with(density) { (visibleRight - visibleLeft).toDp() })
-                        .height(24.dp)
-                        .zIndex(20f)
-                        .padding(horizontal = 2.dp)
-                        .clip(RoundedCornerShape(8.dp))
-                        .background(Color(draft.color).copy(alpha = 0.14f))
-                        .border(1.5.dp, Color(draft.color), RoundedCornerShape(8.dp))
-                        .clickable(
-                            interactionSource = remember { MutableInteractionSource() },
-                            indication = null,
-                            onClick = onDraftTap,
-                        ),
-                )
-            }
-        }
-    }
-}
 
-@Composable
-private fun AllDayContinuationChip(
-    segment: AllDayContinuationSegment,
-    fromPreviousAlpha: Float,
-    toNextAlpha: Float,
-    sideReachFraction: Float,
-    modifier: Modifier,
-) {
-    val darkPalette = CurrentDarkPalette
-    val eventVisuals = remember(segment.item.event?.status, segment.item.event?.manualColor, segment.item.event?.color, darkPalette) {
-        segment.item.event?.cardVisuals(darkPalette = darkPalette)
-    }
-    val base = eventVisuals?.background ?: Color(segment.item.color)
-    val itemAlpha = if (segment.item.completed) 0.42f else 1f
-    val leftAlpha = itemAlpha * fromPreviousAlpha
-    val rightAlpha = itemAlpha * toNextAlpha
-    val leftReach = (sideReachFraction * fromPreviousAlpha.coerceIn(0f, 1f)).coerceIn(0f, 1f)
-    val rightReach = (sideReachFraction * toNextAlpha.coerceIn(0f, 1f)).coerceIn(0f, 1f)
-    val stops = when {
-        segment.fromPrevious && segment.toNext -> arrayOf(
-            0f to base.copy(alpha = leftAlpha),
-            (leftReach * 0.42f) to base.copy(alpha = leftAlpha),
-            leftReach to base.copy(alpha = 0f),
-            (1f - rightReach) to base.copy(alpha = 0f),
-            (1f - rightReach * 0.42f) to base.copy(alpha = rightAlpha),
-            1f to base.copy(alpha = rightAlpha),
-        )
-        segment.fromPrevious -> arrayOf(
-            0f to base.copy(alpha = leftAlpha),
-            (leftReach * 0.42f) to base.copy(alpha = leftAlpha),
-            leftReach to base.copy(alpha = 0f),
-            1f to base.copy(alpha = 0f),
-        )
-        segment.toNext -> arrayOf(
-            0f to base.copy(alpha = 0f),
-            (1f - rightReach) to base.copy(alpha = 0f),
-            (1f - rightReach * 0.42f) to base.copy(alpha = rightAlpha),
-            1f to base.copy(alpha = rightAlpha),
-        )
-        else -> arrayOf(
-            0f to base.copy(alpha = 0f),
-            1f to base.copy(alpha = 0f),
-        )
-    }
-    Box(
-        modifier = modifier
-            .background(Brush.horizontalGradient(*stops)),
-    )
-}
 
-@Composable
-private fun AllDayOverflowChip(
-    modifier: Modifier,
-    hiddenItems: List<AllDayOverlayItem>,
-    onClick: () -> Unit,
-) {
-    val shape = RoundedCornerShape(8.dp)
-    val frontColor = if (MaterialTheme.colorScheme.background.isDark()) Color(0xFF7E8A96) else Color(0xFFA4AFBA)
-    val rearColor = if (MaterialTheme.colorScheme.background.isDark()) Color(0xFF687683) else Color(0xFFB4BEC8)
-    val stackColors = listOf(
-        rearColor,
-        rearColor.copy(alpha = 0.92f),
-    )
-    val textColor = if (frontColor.isDark()) Color.White else Color(0xFF1C1A18)
-    Box(
-        modifier = modifier
-            .clickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = null,
-                onClick = onClick,
-            ),
-        contentAlignment = Alignment.Center,
-    ) {
-        stackColors.forEachIndexed { index, color ->
-            Box(
-                Modifier
-                    .fillMaxWidth()
-                    .height(18.dp)
-                    .align(Alignment.TopCenter)
-                    .offset(y = (3 + index * 2).dp)
-                    .padding(horizontal = (4 + index * 3).dp)
-                    .clip(shape)
-                    .background(color),
-            )
-        }
-        Box(
-            Modifier
-                .fillMaxWidth()
-                .height(20.dp)
-                .align(Alignment.TopCenter)
-                .clip(shape)
-                .background(frontColor),
-            contentAlignment = Alignment.Center,
-        ) {
-            Text(
-                "•••",
-                color = textColor,
-                fontSize = 13.sp,
-                lineHeight = 14.sp,
-                fontWeight = FontWeight.SemiBold,
-                maxLines = 1,
-                softWrap = false,
-                overflow = TextOverflow.Clip,
-            )
-        }
-    }
-}
-
-@Composable
-private fun AllDayViewportChip(
-    item: AllDayOverlayItem,
-    modifier: Modifier,
-    chipLeftPx: Float,
-    chipTopDp: Dp,
-    allDayHeight: Dp,
-    hourHeightDp: Float,
-    timeScrollPx: Int,
-    anchorPage: Int,
-    anchorOffsetPx: Float,
-    dayStepPx: Float,
-    defaultDurationMinutes: Int,
-    shape: RoundedCornerShape = RoundedCornerShape(8.dp),
-    onTaskStatusChanged: (String, String) -> Unit,
-    onDetail: (DetailSheet) -> Unit,
-    onMoveToTimed: (AllDayOverlayItem, LocalDate, LocalTime, LocalTime) -> Unit,
-    onMoveToAllDay: (AllDayOverlayItem, LocalDate) -> Unit,
-    onDragStateChanged: (Boolean) -> Unit,
-) {
-    val darkPalette = CurrentDarkPalette
-    val eventVisuals = remember(item.event?.status, item.event?.manualColor, item.event?.color, darkPalette) {
-        item.event?.cardVisuals(darkPalette = darkPalette)
-    }
-    val color = eventVisuals?.background ?: Color(item.color)
-    val textColor = eventVisuals?.contentColor ?: if (color.isDark()) Color.White else Color(0xFF1C1A18)
-    val eventTextStyle = tentativeReadableTextStyle(item.event?.isTentative() == true)
-    val resourceHref = item.event?.resourceHref ?: item.task?.resourceHref
-    val pendingAlpha = resourceHref?.let { pendingDeleteAlpha(it) } ?: 1f
-    val alpha by animateFloatAsState(
-        targetValue = if (item.completed) 0.48f else 1f,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "allDayViewportChipAlpha",
-    )
-    var lastCompleted by remember(item.task?.uid) { mutableStateOf(item.completed) }
-    var burstKey by remember(item.task?.uid) { mutableStateOf(0) }
-    LaunchedEffect(item.completed) {
-        if (item.completed && !lastCompleted) burstKey++
-        lastCompleted = item.completed
-    }
-    val density = LocalDensity.current
-    var dragX by remember(item.id) { mutableFloatStateOf(0f) }
-    var dragY by remember(item.id) { mutableFloatStateOf(0f) }
-    var isDragging by remember(item.id) { mutableStateOf(false) }
-    var dragPointerOffset by remember(item.id) { mutableStateOf(Offset.Zero) }
-    val chipTopPxForDrag = with(density) { chipTopDp.toPx() }
-    val allDayHeightPxForDrag = with(density) { allDayHeight.toPx() }
-    val hourHeightPxForDrag = with(density) { hourHeightDp.dp.toPx() }
-    val laneTopPx = with(density) { 7.dp.toPx() }
-    val laneStridePx = with(density) { 29.dp.toPx() }
-    val rawFingerX = chipLeftPx + dragPointerOffset.x + dragX
-    val rawFingerY = chipTopPxForDrag + dragPointerOffset.y + dragY
-    val snappedPageDelta = if (dayStepPx > 0f) floor((rawFingerX - anchorOffsetPx) / dayStepPx).toInt() else 0
-    val snappedPage = (anchorPage + snappedPageDelta).coerceIn(0, DayPagerPageCount - 1)
-    val snappedLeftPx = anchorOffsetPx + (snappedPage - anchorPage) * dayStepPx
-    val draggedIntoTimedGrid = rawFingerY >= allDayHeightPxForDrag && hourHeightPxForDrag > 0f && dayStepPx > 0f
-    val snappedDragX = snappedLeftPx - chipLeftPx
-    val snappedDragY = if (draggedIntoTimedGrid) {
-        val duration = defaultDurationMinutes.coerceIn(DraftMinDurationMinutes, 24 * 60 - 1)
-        val gridY = rawFingerY - allDayHeightPxForDrag + timeScrollPx
-        val minuteOffset = ((gridY / hourHeightPxForDrag) * 60f).roundToInt().snapDraftMinute()
-        val startMinute = (DayStartHour * 60 + minuteOffset)
-            .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-        allDayHeightPxForDrag +
-            (((startMinute - DayStartHour * 60) / 60f) * hourHeightPxForDrag) -
-            timeScrollPx -
-            chipTopPxForDrag
-    } else {
-        val lane = if (laneStridePx > 0f) ((rawFingerY - laneTopPx) / laneStridePx).roundToInt().coerceAtLeast(0) else 0
-        (laneTopPx + lane * laneStridePx).coerceIn(0f, (allDayHeightPxForDrag - with(density) { 24.dp.toPx() }).coerceAtLeast(0f)) -
-            chipTopPxForDrag
-    }
-    val displayDragX by animateFloatAsState(
-        targetValue = if (isDragging) snappedDragX else 0f,
-        animationSpec = tween(90, easing = MotionStandard),
-        label = "allDayDraggedSnapX",
-    )
-    val displayDragY by animateFloatAsState(
-        targetValue = if (isDragging) snappedDragY else 0f,
-        animationSpec = tween(90, easing = MotionStandard),
-        label = "allDayDraggedSnapY",
-    )
-    val dragModifier = Modifier.pointerInput(
-        item.id,
-        chipLeftPx,
-        chipTopDp,
-        allDayHeight,
-        hourHeightDp,
-        timeScrollPx,
-        anchorPage,
-        anchorOffsetPx,
-        dayStepPx,
-        defaultDurationMinutes,
-    ) {
-        val chipTopPx = chipTopDp.toPx()
-        val allDayHeightPx = allDayHeight.toPx()
-        val hourHeightPx = hourHeightDp.dp.toPx()
-        detectDragGesturesAfterLongPress(
-            onDragStart = { offset ->
-                dragPointerOffset = offset
-                isDragging = true
-                onDragStateChanged(true)
-            },
-            onDrag = { change, dragAmount ->
-                change.consume()
-                dragX += dragAmount.x
-                dragY += dragAmount.y
-            },
-            onDragCancel = {
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-                onDragStateChanged(false)
-            },
-            onDragEnd = {
-                val fingerX = chipLeftPx + dragPointerOffset.x + dragX
-                val fingerY = chipTopPx + dragPointerOffset.y + dragY
-                val pageDelta = if (dayStepPx > 0f) floor((fingerX - anchorOffsetPx) / dayStepPx).toInt() else 0
-                val targetDate = (anchorPage + pageDelta).coerceIn(0, DayPagerPageCount - 1).toDayDate()
-                if (fingerY >= allDayHeightPx && hourHeightPx > 0f && dayStepPx > 0f) {
-                    val duration = defaultDurationMinutes.coerceIn(DraftMinDurationMinutes, 24 * 60 - 1)
-                    val gridY = fingerY - allDayHeightPx + timeScrollPx
-                    val minuteOffset = ((gridY / hourHeightPx) * 60f).roundToInt().snapDraftMinute()
-                    val startMinute = (DayStartHour * 60 + minuteOffset)
-                        .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-                    onMoveToTimed(item, targetDate, startMinute.toDraftLocalTime(), (startMinute + duration).toDraftLocalTime())
-                } else {
-                    onMoveToAllDay(item, targetDate)
-                }
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-                onDragStateChanged(false)
-            },
-        )
-    }
-    Box(
-        modifier = modifier
-            .zIndex(
-                when {
-                    isDragging -> 80f
-                    item.task != null && !item.completed && taskPriorityIntensity(item.task.priority) > 0f -> 3f
-                    else -> 0f
-                },
-            )
-            .graphicsLayer {
-                translationX = displayDragX
-                translationY = displayDragY
-            }
-            .alpha(alpha * pendingAlpha)
-            .then(
-                if (item.task != null && !item.completed) {
-                    Modifier.taskPriorityMotion(item.task.priority, color)
-                } else {
-                    Modifier
-                },
-            )
-            .background(color, shape)
-            .then(
-                when {
-                    eventVisuals?.dashedBorder == true && eventVisuals.borderColor != null -> Modifier.dashedBorder(eventVisuals.borderColor, 8.dp)
-                    eventVisuals?.borderColor != null -> Modifier.border(1.dp, eventVisuals.borderColor, shape)
-                    else -> Modifier
-                },
-            )
-            .then(dragModifier),
-    ) {
-        Box(
-            Modifier
-                .matchParentSize()
-                .clip(shape)
-                .clickable {
-                    item.event?.let { onDetail(DetailSheet.Event(it)) }
-                    item.task?.let { onDetail(DetailSheet.Task(it)) }
-                },
-        )
-        TaskCardCompletionBurst(burstKey, color = textColor, modifier = Modifier.matchParentSize())
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 6.dp, vertical = 3.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            if (item.task != null) {
-                TaskStatusCheckbox(
-                    status = item.task.effectiveStatus(),
-                    tint = textColor,
-                    onStatusChange = { onTaskStatusChanged(item.task.resourceHref, it) },
-                    boxSize = 18.dp,
-                    iconSize = 15.dp,
-                )
-                Spacer(Modifier.width(4.dp))
-            }
-            Text(
-                item.title,
-                color = textColor,
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 12.sp,
-                lineHeight = 13.sp,
-                maxLines = 1,
-                softWrap = false,
-                overflow = TextOverflow.Clip,
-                textDecoration = eventVisuals?.textDecoration,
-                style = eventTextStyle,
-            )
-        }
-        resourceHref?.let { href ->
-            PendingMutationBadge(href, Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp))
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun AllDayArea(
-    day: LocalDate,
-    visibleStartDate: LocalDate,
-    visibleEndDate: LocalDate,
-    events: List<EventEntity>,
-    tasks: List<TaskEntity>,
-    height: Dp,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onDetail: (DetailSheet) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var taskPopupOpen by remember(day) { mutableStateOf(false) }
-    val quietInteraction = remember { MutableInteractionSource() }
-    val allDayEvents = events.filter { it.isAllDayTopItemOn(day) }.take(2)
-    val allDayTasks = tasks.filter { it.isFullDayTaskOn(day) }
-    // Only the single day being morphed into/out of the month grid registers per-item shared
-    // elements, so all-day chips on neighbouring days don't claim stray morph keys.
-    val morphThisDay = day == LocalMorphItemDay.current
-    if (allDayEvents.isEmpty() && allDayTasks.isEmpty()) {
-        Spacer(
-            modifier
-                .height(height)
-                .fillMaxWidth()
-                .clickable(interactionSource = quietInteraction, indication = null, onClick = {}),
-        )
-        return
-    }
-    if (taskPopupOpen) {
-        KgsModalBottomSheet(
-            onDismissRequest = { taskPopupOpen = false },
-            initialContentHeight = (116 + allDayTasks.size * 58).dp,
-        ) {
-            DayTasksSheet(
-                day = day,
-                tasks = allDayTasks,
-                onTaskStatusChanged = onTaskStatusChanged,
-                onTaskClick = { task ->
-                    taskPopupOpen = false
-                    onDetail(DetailSheet.Task(task))
-                },
-            )
-        }
-    }
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(height)
-            .verticalClipAllowHorizontalOverflow()
-            .clickable(interactionSource = quietInteraction, indication = null, onClick = {})
-            .padding(top = 7.dp, bottom = 8.dp),
-        verticalArrangement = Arrangement.spacedBy(5.dp),
-    ) {
-        allDayEvents.forEach {
-            val starts = it.allDayTopStartDate() ?: day
-            val showTitle = day == maxOf(starts, visibleStartDate)
-            AllDayChip(
-                title = if (showTitle) it.title else "",
-                color = Color(it.displayColor()),
-                event = it,
-                continuesFromPrevious = it.isAllDayTopItemOn(day.minusDays(1)) && day > visibleStartDate,
-                continuesToNext = it.isAllDayTopItemOn(day.plusDays(1)) && day < visibleEndDate,
-                onClick = { onDetail(DetailSheet.Event(it)) },
-                morphDay = morphThisDay,
-            )
-        }
-        if (allDayTasks.isNotEmpty()) {
-            val singleTask = allDayTasks.singleOrNull()
-            val showSingleTaskTitle = singleTask?.let { task ->
-                val starts = task.startAtMillis?.toDate() ?: task.dueAtMillis?.toDate() ?: day
-                day == maxOf(starts, visibleStartDate)
-            } ?: true
-            AllDayTaskChip(
-                title = if (allDayTasks.size == 1) {
-                    if (showSingleTaskTitle) allDayTasks.first().title else ""
-                } else {
-                    stringResource(R.string.task_count, allDayTasks.size)
-                },
-                color = Color(allDayTasks.first().color),
-                completed = allDayTasks.size == 1 && allDayTasks.first().isInactive(),
-                priority = singleTask?.priority,
-                continuesFromPrevious = singleTask?.isFullDayTaskOn(day.minusDays(1)) == true && day > visibleStartDate,
-                continuesToNext = singleTask?.isFullDayTaskOn(day.plusDays(1)) == true && day < visibleEndDate,
-                status = singleTask?.effectiveStatus(),
-                resourceHref = singleTask?.resourceHref,
-                onStatusChange = if (singleTask != null) {
-                    { status -> onTaskStatusChanged(singleTask.resourceHref, status) }
-                } else {
-                    null
-                },
-                onClick = {
-                    if (allDayTasks.size == 1) {
-                        onDetail(DetailSheet.Task(allDayTasks.first()))
-                    } else {
-                        taskPopupOpen = true
-                    }
-                },
-                morphUid = if (morphThisDay && allDayTasks.size == 1) allDayTasks.first().resourceHref else null,
-            )
-        }
-    }
-}
-
-private fun Modifier.verticalPinchZoom(
-    topOffsetPx: Float = 0f,
-    viewportHeightPx: Int = 0,
-    onZoom: (Float, Float) -> Unit,
-): Modifier = pointerInput(topOffsetPx, viewportHeightPx, onZoom) {
-    awaitPointerEventScope {
-        var previousSpanY: Float? = null
-        while (true) {
-            val event = awaitPointerEvent()
-            val pressed = event.changes.filter { it.pressed }
-            if (pressed.size < 2) {
-                previousSpanY = null
-                continue
-            }
-
-            val top = pressed.minOf { it.position.y }
-            val bottom = pressed.maxOf { it.position.y }
-            val spanY = (bottom - top).coerceAtLeast(1f)
-            val centroidY = pressed.sumOf { it.position.y.toDouble() }.toFloat() / pressed.size
-            val previous = previousSpanY
-            if (previous != null) {
-                val zoom = spanY / previous
-                if (zoom.isFinite() && zoom > 0f && abs(zoom - 1f) > 0.01f) {
-                    val gridCentroidY = (centroidY - topOffsetPx).let { y ->
-                        if (viewportHeightPx > 0) y.coerceIn(0f, viewportHeightPx.toFloat()) else y.coerceAtLeast(0f)
-                    }
-                    onZoom(zoom, gridCentroidY)
-                    pressed.forEach { it.consume() }
-                }
-            }
-            previousSpanY = spanY
-        }
-    }
-}
-
-@Composable
-private fun TimeSidebar(hours: IntRange, hourHeightDp: Float, onZoom: (Float, Float) -> Unit) {
-    val currentOnZoom by rememberUpdatedState(onZoom)
-    val hourCount = hours.count()
-    val textColor = WarmInk
-    Box(
-        modifier = Modifier
-            .width(TimeSidebarWidth)
-            .height((hourHeightDp * hourCount).dp)
-            .pointerInput(Unit) {
-                detectTransformGestures { centroid, _, zoom, _ ->
-                    currentOnZoom(zoom, centroid.y)
-                }
-            },
-    ) {
-        Canvas(Modifier.matchParentSize()) {
-            val rowHeight = hourHeightDp.dp.toPx()
-            val paint = Paint(Paint.ANTI_ALIAS_FLAG).apply {
-                color = textColor.toArgb()
-                textSize = 10.sp.toPx()
-                textAlign = Paint.Align.CENTER
-                isFakeBoldText = true
-            }
-            val baselineOffset = 11.sp.toPx()
-            hours.forEachIndexed { index, hour ->
-                drawContext.canvas.nativeCanvas.drawText(
-                    "%02d:00".format(hour),
-                    size.width / 2f,
-                    index * rowHeight + baselineOffset,
-                    paint,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun DayTimedColumn(
-    day: LocalDate,
-    hourHeightDp: Float,
-    dayWidthPx: Float,
-    taskColorMode: TaskColorMode,
-    events: List<EventEntity>,
-    tasks: List<TaskEntity>,
-    draftEvent: DraftEventSelection?,
-    onDraftEventChanged: (DraftEventSelection) -> Unit,
-    onDraftInteraction: () -> Unit,
-    onDraftTap: () -> Unit,
-    timeScroll: androidx.compose.foundation.ScrollState,
-    gridViewportHeightPx: Int,
-    onSlotSelected: (LocalDate, LocalTime) -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEventMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onTaskMoved: (String, Long, LocalDate, LocalTime, LocalTime) -> Unit,
-    onEventMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onTaskMovedAllDay: (String, Long, LocalDate) -> Unit,
-    onDetail: (DetailSheet) -> Unit,
-    modifier: Modifier = Modifier,
-    drawGrid: Boolean = true,
-) {
-    val hourCount = DayEndHour - DayStartHour + 1
-    val gridColor = WarmGrid
-    val timedItems = remember(day, events, tasks, hourHeightDp) {
-        layoutTimedItemsForDay(
-            day = day,
-            hourHeightDp = hourHeightDp,
-            events = events.filter { !it.allDay },
-            tasks = tasks,
-        )
-    }
-    Box(modifier = modifier) {
-        if (drawGrid) {
-            Box(
-                Modifier
-                    .matchParentSize()
-                    .drawBehind {
-                        val rowHeight = hourHeightDp.dp.toPx()
-                        val cellHeight = (rowHeight - HourCellGap.toPx()).coerceAtLeast(0f)
-                        val cellInset = HourCellGap.toPx() / 2f
-                        val radius = 10.dp.toPx()
-                        repeat(hourCount) { index ->
-                            val top = index * rowHeight + cellInset
-                            val cellSize = Size(width = size.width, height = cellHeight)
-                            val cornerRadius = CornerRadius(radius, radius)
-                            drawRoundRect(
-                                color = gridColor,
-                                topLeft = Offset(0f, top),
-                                size = cellSize,
-                                cornerRadius = cornerRadius,
-                            )
-                        }
-                    },
-            )
-        }
-        timedItems.forEach { layout ->
-            when (val item = layout.item) {
-                is TimedCalendarItem.EventItem -> TimedEventBlock(
-                    event = item.event,
-                    placement = layout.placement,
-                    lane = layout.lane,
-                    laneCount = layout.laneCount,
-                    day = day,
-                    hourHeightDp = hourHeightDp,
-                    color = item.event.displayColor(),
-                    dayWidthPx = dayWidthPx,
-                    onMove = { targetDate, start, end -> onEventMoved(item.event.resourceHref, item.event.startsAtMillis, targetDate, start, end) },
-                    onMoveAllDay = { targetDate -> onEventMovedAllDay(item.event.resourceHref, item.event.startsAtMillis, targetDate) },
-                    onClick = { onDetail(DetailSheet.Event(item.event)) },
-                )
-                is TimedCalendarItem.TaskItem -> TimedTaskBlock(
-                    task = item.task,
-                    placement = layout.placement,
-                    lane = layout.lane,
-                    laneCount = layout.laneCount,
-                    day = day,
-                    hourHeightDp = hourHeightDp,
-                    color = item.task.displayColor(taskColorMode),
-                    dayWidthPx = dayWidthPx,
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onMove = { targetDate, start, end ->
-                        onTaskMoved(
-                            item.task.resourceHref,
-                            item.task.startAtMillis ?: item.task.dueAtMillis ?: System.currentTimeMillis(),
-                            targetDate,
-                            start,
-                            end,
-                        )
-                    },
-                    onMoveAllDay = { targetDate ->
-                        onTaskMovedAllDay(
-                            item.task.resourceHref,
-                            item.task.startAtMillis ?: item.task.dueAtMillis ?: System.currentTimeMillis(),
-                            targetDate,
-                        )
-                    },
-                    onClick = { onDetail(DetailSheet.Task(item.task)) },
-                )
-            }
-        }
-        draftEvent?.takeIf { !it.allDay }?.let { draft ->
-            DraftEventWireframe(
-                draft = EditorSchedulePreview(draft.date, draft.start, draft.end, draft.allDay),
-                color = draft.color,
-                hourHeightDp = hourHeightDp,
-                dayWidthPx = dayWidthPx,
-                onDraftChanged = { changed ->
-                    onDraftEventChanged(
-                        DraftEventSelection(changed.date, changed.start, changed.end, draft.color, changed.allDay),
-                    )
-                },
-                onInteraction = onDraftInteraction,
-                onTap = onDraftTap,
-                timeScroll = timeScroll,
-                gridViewportHeightPx = gridViewportHeightPx,
-            )
-        }
-    }
-}
-
-@Composable
-private fun FadingTimedText(
-    text: String,
-    color: Color,
-    fontSize: TextUnit,
-    lineHeight: TextUnit,
-    modifier: Modifier = Modifier,
-    fontWeight: FontWeight = FontWeight.Normal,
-    maxLines: Int = 1,
-    textScale: Float = 1f,
-    textDecoration: TextDecoration? = null,
-    shadow: Shadow? = null,
-) {
-    // Single-line text must not wrap; otherwise a title like "Klavier Maximilian"
-    // wraps after the first word and the wrapped line gets clipped by maxLines=1,
-    // leaving only "Klavier" visible. With softWrap disabled the overflow runs
-    // off the right edge where the gradient below fades it out.
-    val softWrap = maxLines > 1
-    // Scale by adjusting font/line-height directly rather than via graphicsLayer.
-    // graphicsLayer scaling with transformOrigin = (0,0) shrinks the rendered text
-    // toward the top-left, leaving the right side of the layout visually empty —
-    // which made the gradient look like it started far from the card's edge.
-    val scaledFontSize = (fontSize.value * textScale).sp
-    val scaledLineHeight = (lineHeight.value * textScale).sp
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .rightEdgeFadeMask(),
-    ) {
-        Text(
-            text = text,
-            color = color,
-            fontSize = scaledFontSize,
-            lineHeight = scaledLineHeight,
-            fontWeight = fontWeight,
-            textDecoration = textDecoration,
-            style = TextStyle(shadow = shadow),
-            maxLines = maxLines,
-            softWrap = softWrap,
-            overflow = TextOverflow.Clip,
-            modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-private fun Modifier.rightEdgeFadeMask(edgeWidth: Dp = 22.dp): Modifier =
-    graphicsLayer { compositingStrategy = CompositingStrategy.Offscreen }
-        .drawWithContent {
-            drawContent()
-            val edge = edgeWidth.toPx().coerceAtMost(size.width)
-            drawRect(
-                brush = Brush.horizontalGradient(
-                    0.00f to Color.Black,
-                    0.55f to Color.Black.copy(alpha = 0.95f),
-                    0.78f to Color.Black.copy(alpha = 0.65f),
-                    0.92f to Color.Black.copy(alpha = 0.22f),
-                    1.00f to Color.Transparent,
-                    startX = size.width - edge,
-                    endX = size.width,
-                ),
-                topLeft = Offset(size.width - edge, 0f),
-                size = Size(edge, size.height),
-                blendMode = BlendMode.DstIn,
-            )
-        }
-
-private fun Modifier.bottomEdgeFadeMask(edgeHeight: Dp = 8.dp): Modifier =
-    graphicsLayer { compositingStrategy = CompositingStrategy.Offscreen }
-        .drawWithContent {
-            drawContent()
-            val edge = edgeHeight.toPx().coerceAtMost(size.height)
-            drawRect(
-                brush = Brush.verticalGradient(
-                    0f to Color.Black,
-                    1f to Color.Transparent,
-                    startY = size.height - edge,
-                    endY = size.height,
-                ),
-                topLeft = Offset(0f, size.height - edge),
-                size = Size(size.width, edge),
-                blendMode = BlendMode.DstIn,
-            )
-        }
-
-@Composable
-private fun tentativeReadableTextStyle(enabled: Boolean): TextStyle =
-    if (enabled) {
-        val backdrop = if (MaterialTheme.colorScheme.background.isDark()) Color.Black else Color.White
-        TextStyle(
-            shadow = Shadow(
-                color = backdrop.copy(alpha = 0.88f),
-                offset = Offset(0f, 1f),
-                blurRadius = 3.2f,
-            ),
-        )
-    } else {
-        TextStyle.Default
-    }
-
-private enum class ParticipantDisplayMode { Hidden, Compact, Full }
-
-private data class ParticipantDisplayTarget(
-    val mode: ParticipantDisplayMode,
-    val fullLimit: Int,
-)
-
-@Composable
-private fun TimedEventBlock(
-    event: EventEntity,
-    placement: TimedPlacement,
-    lane: Int,
-    laneCount: Int,
-    day: LocalDate,
-    hourHeightDp: Float,
-    color: Int,
-    dayWidthPx: Float,
-    onMove: (LocalDate, LocalTime, LocalTime) -> Unit,
-    onMoveAllDay: (LocalDate) -> Unit,
-    onClick: () -> Unit,
-) {
-    val darkPalette = CurrentDarkPalette
-    val visuals = remember(event.status, event.manualColor, event.color, color, darkPalette) {
-        event.cardVisuals(darkPalette = darkPalette)
-    }
-    val background = visuals.background
-    val textColor = visuals.contentColor
-    val pendingAlpha = pendingDeleteAlpha(event.resourceHref)
-    val attendees = remember(event.attendeesJson) { event.attendeesJson.toCalendarParticipants() }
-    var dragX by remember(event.resourceHref, event.startsAtMillis) { mutableStateOf(0f) }
-    var dragY by remember(event.resourceHref, event.startsAtMillis) { mutableStateOf(0f) }
-    var isDragging by remember(event.resourceHref, event.startsAtMillis) { mutableStateOf(false) }
-    val dragReporter = LocalTimedDragReporter.current
-    LaunchedEffect(isDragging) { dragReporter(day, isDragging) }
-    DisposableEffect(event.resourceHref, event.startsAtMillis) {
-        onDispose { dragReporter(day, false) }
-    }
-    val density = LocalDensity.current
-    val hourHeightPxForDrag = with(density) { hourHeightDp.dp.toPx() }
-    val dayStepPx = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-    val duration = (placement.endMinute - placement.startMinute).coerceAtLeast(DraftMinDurationMinutes)
-    val snappedDayDelta = if (dayStepPx > 0f) (dragX / dayStepPx).roundToInt() else 0
-    val snappedMinuteDelta = if (hourHeightPxForDrag > 0f) {
-        ((dragY / hourHeightPxForDrag) * 60f).roundToInt().snapDraftMinute()
-    } else {
-        0
-    }
-    val snappedStart = (placement.startMinute + snappedMinuteDelta)
-        .snapDraftMinute()
-        .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-    val snappedDragYPx = with(density) { ((snappedStart - placement.startMinute) / 60f * hourHeightDp).dp.toPx() }
-    val animatedDragX by animateFloatAsState(
-        targetValue = if (isDragging) snappedDayDelta * dayStepPx else 0f,
-        animationSpec = tween(110, easing = MotionStandard),
-        label = "eventDragX",
-    )
-    val animatedDragY by animateFloatAsState(
-        targetValue = if (isDragging) snappedDragYPx else 0f,
-        animationSpec = tween(110, easing = MotionStandard),
-        label = "eventDragY",
-    )
-    val titleScale = ((placement.heightDp - 13f) / 20f).coerceIn(0f, 1f)
-    val compactVerticalPadding = (1.5f + titleScale * 1.5f).dp
-    val compactTitleScale = 0.9f + titleScale * 0.1f
-    val showTitle = placement.heightDp >= 13f
-    val titleMaxLines = if (placement.heightDp >= 43f) 2 else 1
-    val displayLocation = remember(event.location, event.locationMapVerified) {
-        event.location
-            ?.cardLocationText(event.locationMapVerified)
-            .orEmpty()
-    }
-    val showLocation = placement.heightDp >= 40f && displayLocation.isNotBlank()
-    val locationMaxLines = when {
-        placement.heightDp >= 86f -> 4
-        placement.heightDp >= 68f -> 3
-        placement.heightDp >= 52f -> 2
-        else -> 1
-    }
-    val participantMode = when {
-        attendees.isEmpty() || placement.heightDp < 5f -> ParticipantDisplayMode.Hidden
-        placement.heightDp >= 72f -> ParticipantDisplayMode.Full
-        else -> ParticipantDisplayMode.Compact
-    }
-    val compactParticipantScale = ((placement.heightDp - 5f) / 18f).coerceIn(0.34f, 1f)
-    val participantLimit = if (displayLocation.isNotBlank() && placement.heightDp < 88f) 2 else 6
-    val participantTarget = ParticipantDisplayTarget(
-        mode = participantMode,
-        fullLimit = if (participantMode == ParticipantDisplayMode.Full) participantLimit else 0,
-    )
-    val fadeTextBottom = titleMaxLines > 1 || showLocation
-    val textShadow = if (event.isTentative()) {
-        Shadow(
-            color = MaterialTheme.colorScheme.surface.copy(alpha = 0.92f),
-            offset = Offset(0f, 1f),
-            blurRadius = 3.2f,
-        )
-    } else {
-        null
-    }
-    val dragModifier = Modifier.pointerInput(event.resourceHref, event.startsAtMillis, placement, hourHeightDp, dayWidthPx) {
-        val hourHeightPx = with(density) { hourHeightDp.dp.toPx() }
-        val dayStep = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-        detectDragGesturesAfterLongPress(
-            onDragStart = {
-                isDragging = true
-            },
-            onDrag = { change, dragAmount ->
-                change.consume()
-                dragX += dragAmount.x
-                dragY += dragAmount.y
-            },
-            onDragCancel = {
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-            },
-            onDragEnd = {
-                val dayDelta = if (dayStep > 0f) (dragX / dayStep).roundToInt() else 0
-                val minuteDelta = ((dragY / hourHeightPx) * 60f).roundToInt().snapDraftMinute()
-                val nextStart = (placement.startMinute + minuteDelta)
-                    .snapDraftMinute()
-                    .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-                val targetDate = day.plusDays(dayDelta.toLong())
-                val movedTopPx = with(density) { placement.topDp.dp.toPx() } + dragY
-                if (movedTopPx < with(density) { (-16).dp.toPx() }) {
-                    onMoveAllDay(targetDate)
-                } else {
-                    onMove(targetDate, nextStart.toDraftLocalTime(), (nextStart + duration).toDraftLocalTime())
-                }
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-            },
-        )
-    }
-    Row(
-        modifier = Modifier
-            .offset {
-                IntOffset(
-                    animatedDragX.roundToInt(),
-                    with(density) { (placement.topDp.dp.toPx() + animatedDragY).roundToInt() },
-                )
-            }
-            .zIndex(if (isDragging) 20f else 1f)
-            .fillMaxWidth()
-            .height(placement.heightDp.dp),
-    ) {
-        if (lane > 0) Spacer(Modifier.weight(lane.toFloat()))
-        val shape = RoundedCornerShape(10.dp)
-        Box(
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight()
-                // Option B: morph this card out of / into its month-cell pill (matched by uid).
-                .morphBounds(
-                    key = "item-${event.resourceHref}",
-                    enabled = EnablePerItemMorph && day == LocalMorphItemDay.current,
-                    enter = MorphItemCardEnter,
-                    exit = MorphItemCardExit,
-                    boundsTransform = MorphItemBoundsTransform,
-                )
-                .padding(vertical = 1.dp)
-                .alpha(pendingAlpha)
-                .background(background, shape)
-                .then(dragModifier),
-        ) {
-            Box(
-                Modifier
-                    .matchParentSize()
-                    .clip(shape)
-                    .clickable(onClick = onClick),
-            )
-            Box(Modifier.matchParentSize()) {
-                if (showTitle) {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .then(if (fadeTextBottom) Modifier.bottomEdgeFadeMask() else Modifier)
-                            .padding(start = 6.dp, end = 0.dp, top = compactVerticalPadding, bottom = 0.dp),
-                        verticalArrangement = Arrangement.Top,
-                    ) {
-                        FadingTimedText(
-                            text = event.title,
-                            color = textColor,
-                            fontSize = 12.sp,
-                            lineHeight = 13.sp,
-                            fontWeight = FontWeight.Medium,
-                            maxLines = titleMaxLines,
-                            textScale = compactTitleScale,
-                            textDecoration = visuals.textDecoration,
-                            shadow = textShadow,
-                        )
-                        if (showLocation) {
-                            FadingTimedText(
-                                text = displayLocation,
-                                color = textColor.copy(alpha = 0.86f),
-                                fontSize = 11.sp,
-                                lineHeight = 12.sp,
-                                maxLines = locationMaxLines,
-                                textDecoration = visuals.textDecoration,
-                                shadow = textShadow,
-                            )
-                        }
-                    }
-                }
-            }
-            AnimatedContent(
-                targetState = participantTarget,
-                modifier = Modifier
-                    .align(Alignment.BottomEnd)
-                    .fillMaxWidth(),
-                transitionSpec = {
-                    val shrinking = initialState.mode == ParticipantDisplayMode.Full &&
-                        (targetState.mode != ParticipantDisplayMode.Full || targetState.fullLimit < initialState.fullLimit)
-                    val entering = if (shrinking) {
-                        fadeIn(tween(120, delayMillis = 90, easing = MotionStandard)) +
-                            scaleIn(tween(120, delayMillis = 90, easing = MotionStandard), initialScale = 0.9f)
-                    } else {
-                        fadeIn(tween(150, easing = MotionStandard)) +
-                            scaleIn(tween(150, easing = MotionStandard), initialScale = 0.86f)
-                    }
-                    val exiting = if (shrinking) {
-                        fadeOut(tween(90, easing = MotionStandard)) +
-                            scaleOut(tween(90, easing = MotionStandard), targetScale = 0.92f)
-                    } else {
-                        fadeOut(tween(110, easing = MotionStandard)) +
-                            scaleOut(tween(110, easing = MotionStandard), targetScale = 0.86f)
-                    }
-                    entering.togetherWith(exiting).using(SizeTransform(clip = false))
-                },
-                label = "participantStackMorph",
-            ) { target ->
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .graphicsLayer { clip = false },
-                    contentAlignment = Alignment.BottomEnd,
-                ) {
-                    when (target.mode) {
-                        ParticipantDisplayMode.Full -> EventParticipantStack(
-                            attendees = attendees,
-                            eventColor = visuals.baseColor,
-                            contentColor = textColor,
-                            modifier = Modifier
-                                .padding(end = 4.dp, bottom = 2.dp)
-                                .fillMaxWidth(0.92f),
-                            circleSize = 22.dp,
-                            maxVisible = target.fullLimit,
-                        )
-
-                        ParticipantDisplayMode.Compact -> CompactParticipantStatusDots(
-                            attendees = attendees,
-                            scale = compactParticipantScale,
-                            modifier = Modifier.padding(end = 3.dp),
-                        )
-
-                        ParticipantDisplayMode.Hidden -> Spacer(Modifier.size(0.dp))
-                    }
-                }
-            }
-            Box(
-                Modifier
-                    .matchParentSize()
-                    .then(
-                        when {
-                            visuals.dashedBorder && visuals.borderColor != null -> Modifier.dashedBorder(visuals.borderColor, 10.dp)
-                            visuals.borderColor != null -> Modifier.border(1.dp, visuals.borderColor, shape)
-                            else -> Modifier
-                        },
-                    ),
-            )
-            PendingMutationBadge(
-                resourceHref = event.resourceHref,
-                modifier = Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp),
-            )
-        }
-        if (laneCount - lane - 1 > 0) Spacer(Modifier.weight((laneCount - lane - 1).toFloat()))
-    }
-}
-
-@Composable
-private fun TimedTaskBlock(
-    task: TaskEntity,
-    placement: TimedPlacement,
-    lane: Int,
-    laneCount: Int,
-    day: LocalDate,
-    hourHeightDp: Float,
-    color: Int,
-    dayWidthPx: Float,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onMove: (LocalDate, LocalTime, LocalTime) -> Unit,
-    onMoveAllDay: (LocalDate) -> Unit,
-    onClick: () -> Unit,
-) {
-    val background = remember(color) { Color(color) }
-    val textColor = remember(color) { if (background.isDark()) Color.White else Color(0xFF1C1A18) }
-    val pendingAlpha = pendingDeleteAlpha(task.resourceHref)
-    var lastCompleted by remember(task.resourceHref) { mutableStateOf(task.isCompleted) }
-    var burstKey by remember(task.resourceHref) { mutableStateOf(0) }
-    LaunchedEffect(task.isCompleted) {
-        if (task.isCompleted && !lastCompleted) burstKey++
-        lastCompleted = task.isCompleted
-    }
-    val inactive = task.isInactive()
-    val taskAlpha by animateFloatAsState(
-        targetValue = if (inactive) 0.48f else 1f,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "timedTaskAlpha",
-    )
-    var dragX by remember(task.resourceHref) { mutableStateOf(0f) }
-    var dragY by remember(task.resourceHref) { mutableStateOf(0f) }
-    var isDragging by remember(task.resourceHref) { mutableStateOf(false) }
-    val dragReporter = LocalTimedDragReporter.current
-    LaunchedEffect(isDragging) { dragReporter(day, isDragging) }
-    DisposableEffect(task.resourceHref) {
-        onDispose { dragReporter(day, false) }
-    }
-    val density = LocalDensity.current
-    val hourHeightPxForDrag = with(density) { hourHeightDp.dp.toPx() }
-    val dayStepPx = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-    val duration = (placement.endMinute - placement.startMinute).coerceAtLeast(DraftMinDurationMinutes)
-    val snappedDayDelta = if (dayStepPx > 0f) (dragX / dayStepPx).roundToInt() else 0
-    val snappedMinuteDelta = if (hourHeightPxForDrag > 0f) {
-        ((dragY / hourHeightPxForDrag) * 60f).roundToInt().snapDraftMinute()
-    } else {
-        0
-    }
-    val snappedStart = (placement.startMinute + snappedMinuteDelta)
-        .snapDraftMinute()
-        .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-    val snappedDragYPx = with(density) { ((snappedStart - placement.startMinute) / 60f * hourHeightDp).dp.toPx() }
-    val animatedDragX by animateFloatAsState(
-        targetValue = if (isDragging) snappedDayDelta * dayStepPx else 0f,
-        animationSpec = tween(110, easing = MotionStandard),
-        label = "taskDragX",
-    )
-    val animatedDragY by animateFloatAsState(
-        targetValue = if (isDragging) snappedDragYPx else 0f,
-        animationSpec = tween(110, easing = MotionStandard),
-        label = "taskDragY",
-    )
-    val titleScale = ((placement.heightDp - 18f) / 18f).coerceIn(0f, 1f)
-    val compactVerticalPadding = (1.5f + titleScale * 1.5f).dp
-    val compactTitleScale = 0.9f + titleScale * 0.1f
-    val showTitle = placement.heightDp >= 18f
-    val titleMaxLines = if (placement.heightDp >= 43f) 2 else 1
-    val displayLocation = remember(task.location, task.locationMapVerified) {
-        task.location
-            ?.cardLocationText(task.locationMapVerified)
-            .orEmpty()
-    }
-    val showLocation = placement.heightDp >= 42f && displayLocation.isNotBlank()
-    val locationMaxLines = when {
-        placement.heightDp >= 86f -> 4
-        placement.heightDp >= 68f -> 3
-        placement.heightDp >= 52f -> 2
-        else -> 1
-    }
-    val checkboxBoxSize = min(18f, (placement.heightDp - 2f).coerceAtLeast(15f))
-    val checkboxIconSize = min(15f, checkboxBoxSize)
-    val checkboxTopOffset = min(
-        compactVerticalPadding.value,
-        max(0f, ((placement.heightDp - 2f) - checkboxBoxSize) / 2f),
-    ).dp
-    val dragModifier = Modifier.pointerInput(task.resourceHref, placement, hourHeightDp, dayWidthPx) {
-        val hourHeightPx = with(density) { hourHeightDp.dp.toPx() }
-        val dayStep = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-        detectDragGesturesAfterLongPress(
-            onDragStart = {
-                isDragging = true
-            },
-            onDrag = { change, dragAmount ->
-                change.consume()
-                dragX += dragAmount.x
-                dragY += dragAmount.y
-            },
-            onDragCancel = {
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-            },
-            onDragEnd = {
-                val dayDelta = if (dayStep > 0f) (dragX / dayStep).roundToInt() else 0
-                val minuteDelta = ((dragY / hourHeightPx) * 60f).roundToInt().snapDraftMinute()
-                val nextStart = (placement.startMinute + minuteDelta)
-                    .snapDraftMinute()
-                    .coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - duration)
-                val targetDate = day.plusDays(dayDelta.toLong())
-                val movedTopPx = with(density) { placement.topDp.dp.toPx() } + dragY
-                if (movedTopPx < with(density) { (-16).dp.toPx() }) {
-                    onMoveAllDay(targetDate)
-                } else {
-                    onMove(targetDate, nextStart.toDraftLocalTime(), (nextStart + duration).toDraftLocalTime())
-                }
-                isDragging = false
-                dragX = 0f
-                dragY = 0f
-            },
-        )
-    }
-    Row(
-        modifier = Modifier
-            .offset {
-                IntOffset(
-                    animatedDragX.roundToInt(),
-                    with(density) { (placement.topDp.dp.toPx() + animatedDragY).roundToInt() },
-                )
-            }
-            .zIndex(if (isDragging) 20f else if (!inactive && taskPriorityIntensity(task.priority) > 0f) 3f else 1f)
-            .fillMaxWidth()
-            .height(placement.heightDp.dp),
-        verticalAlignment = Alignment.Top,
-    ) {
-        if (lane > 0) Spacer(Modifier.weight(lane.toFloat()))
-        val shape = RoundedCornerShape(10.dp)
-        Box(
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxHeight()
-                // Option B: morph this task card out of / into its month-cell pill (matched by uid).
-                .morphBounds(
-                    key = "item-${task.resourceHref}",
-                    enabled = EnablePerItemMorph && day == LocalMorphItemDay.current,
-                    enter = MorphItemCardEnter,
-                    exit = MorphItemCardExit,
-                    boundsTransform = MorphItemBoundsTransform,
-                )
-                .taskPriorityMotion(if (inactive) null else task.priority, background)
-                .padding(vertical = 1.dp)
-                .alpha(taskAlpha * pendingAlpha)
-                .background(background, shape)
-                .then(dragModifier),
-        ) {
-            TaskCardCompletionBurst(burstKey, color = textColor, modifier = Modifier.matchParentSize())
-            Box(
-                Modifier
-                    .matchParentSize()
-                    .clip(shape)
-                    .clickable(onClick = onClick),
-            )
-            Row(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(start = 5.dp),
-                verticalAlignment = Alignment.Top,
-            ) {
-                TaskStatusCheckbox(
-                    status = task.effectiveStatus(),
-                    tint = textColor,
-                    onStatusChange = { onTaskStatusChanged(task.resourceHref, it) },
-                    boxSize = checkboxBoxSize.dp,
-                    iconSize = checkboxIconSize.dp,
-                    modifier = Modifier.offset(y = checkboxTopOffset),
-                )
-                Spacer(Modifier.width(3.dp))
-                if (showTitle) {
-                    Column(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxHeight()
-                            .then(if (titleMaxLines > 1 || showLocation) Modifier.bottomEdgeFadeMask() else Modifier)
-                            .padding(top = compactVerticalPadding),
-                    ) {
-                        FadingTimedText(
-                            text = task.title,
-                            color = textColor,
-                            maxLines = titleMaxLines,
-                            fontSize = 12.sp,
-                            lineHeight = 13.sp,
-                            fontWeight = FontWeight.Medium,
-                            textScale = compactTitleScale,
-                        )
-                        if (showLocation) {
-                            FadingTimedText(
-                                text = displayLocation,
-                                color = textColor.copy(alpha = 0.86f),
-                                fontSize = 11.sp,
-                                lineHeight = 12.sp,
-                                maxLines = locationMaxLines,
-                            )
-                        }
-                    }
-                }
-            }
-            PendingMutationBadge(
-                resourceHref = task.resourceHref,
-                modifier = Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp),
-            )
-        }
-        if (laneCount - lane - 1 > 0) Spacer(Modifier.weight((laneCount - lane - 1).toFloat()))
-    }
-}
-
-@Composable
-private fun Modifier.taskPriorityMotion(
-    priority: Int?,
-    color: Color,
-    expandHorizontally: Boolean = true,
-): Modifier {
-    if (!PriorityAnimationsEnabled) return this
-    val intensity = taskPriorityIntensity(priority)
-    if (intensity <= 0f) return this
-    val transition = rememberInfiniteTransition(label = "taskPriorityMotion")
-    val pulse by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 1f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(
-                durationMillis = (1050 - (intensity * 420f)).roundToInt().coerceAtLeast(520),
-                easing = MotionStandard,
-            ),
-            repeatMode = RepeatMode.Reverse,
-        ),
-        label = "taskPriorityPulseValue",
-    )
-    val shakePhase by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 1f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(260, easing = LinearEasing),
-            repeatMode = RepeatMode.Restart,
-        ),
-        label = "taskPriorityShakePhase",
-    )
-    val highestPriorityShake = if (priority == 1) cos(shakePhase * PI * 6.0).toFloat() * 1.6f else 0f
-    val bounce = ((pulse - 0.5f) * -2f * intensity)
-    val scale = 1f + intensity * 0.018f * pulse
-    return this
-        .graphicsLayer {
-            translationX = highestPriorityShake
-            translationY = bounce
-            scaleX = if (expandHorizontally) scale else 1f
-            scaleY = scale
-        }
-        .drawBehind {
-            val spread = 8.dp.toPx() * intensity * (0.45f + pulse)
-            val horizontalSpread = if (expandHorizontally) spread else 0f
-            drawRoundRect(
-                color = color.copy(alpha = 0.18f * intensity * (0.45f + 0.55f * pulse)),
-                topLeft = Offset(-horizontalSpread / 2f, -spread / 2f),
-                size = Size(size.width + horizontalSpread, size.height + spread),
-                cornerRadius = CornerRadius(8.dp.toPx() + spread / 3f, 8.dp.toPx() + spread / 3f),
-            )
-        }
-}
-
-@Composable
-internal fun DraftEventWireframe(
-    draft: EditorSchedulePreview,
-    color: Int,
-    hourHeightDp: Float,
-    dayWidthPx: Float,
-    onDraftChanged: (EditorSchedulePreview) -> Unit,
-    onInteraction: () -> Unit,
-    onTap: () -> Unit,
-    timeScroll: androidx.compose.foundation.ScrollState,
-    gridViewportHeightPx: Int,
-) {
-    val density = LocalDensity.current
-    val currentDraft by rememberUpdatedState(draft)
-    val currentOnDraftChanged by rememberUpdatedState(onDraftChanged)
-    val visibleStartMinute = DayStartHour * 60
-    val visibleEndMinute = (DayEndHour + 1) * 60 - 1
-    val startMinute = draft.start.minuteOfDay().coerceIn(visibleStartMinute, visibleEndMinute - DraftMinDurationMinutes)
-    val endMinute = draft.end.minuteOfDay().coerceIn(startMinute + DraftMinDurationMinutes, visibleEndMinute)
-    val topDp = (startMinute - visibleStartMinute) / 60f * hourHeightDp
-    val visualHeightDp = ((endMinute - startMinute) / 60f * hourHeightDp).coerceAtLeast(1f)
-    val touchHeightDp = max(44f, visualHeightDp)
-    val visualTopInsetDp = (touchHeightDp - visualHeightDp) / 2f
-    val shape = RoundedCornerShape(8.dp)
-    val wireframeColor = Color(color)
-    var dragX by remember { mutableFloatStateOf(0f) }
-    var draggingBody by remember { mutableStateOf(false) }
-    val dayStepPx = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-    val snappedDayDelta = if (draggingBody && dayStepPx > 0f) (dragX / dayStepPx).roundToInt() else 0
-    val animatedDragX by animateFloatAsState(
-        targetValue = if (draggingBody) snappedDayDelta * dayStepPx else 0f,
-        animationSpec = tween(110, easing = MotionStandard),
-        label = "draftWireframeDragX",
-    )
-
-    fun Modifier.draftDrag(mode: DraftDragMode): Modifier = pointerInput(hourHeightDp, dayWidthPx, mode, gridViewportHeightPx) {
-        coroutineScope {
-        val hourHeightPx = with(density) { hourHeightDp.dp.toPx() }
-        val edgeThresholdPx = with(density) { 34.dp.toPx() }
-        val dayStep = dayWidthPx + with(density) { DayColumnSpacing.toPx() }
-        var dragStartMinute = 0
-        var dragEndMinute = 0
-        var dragStartDate = currentDraft.date
-        var totalDragPx = 0f
-        var autoScrollDirection = 0f
-        var autoScrollJob: Job? = null
-
-        fun publishDraft(commitDate: Boolean = false) {
-            val deltaMinutes = ((totalDragPx / hourHeightPx) * 60f).roundToInt()
-            val dayDelta = if (mode == DraftDragMode.Move && dayStep > 0f) {
-                (dragX / dayStep).roundToInt()
-            } else {
-                0
-            }
-            val nextDraft = currentDraft.withDraggedMinutes(
-                    mode = mode,
-                    originalStartMinute = dragStartMinute,
-                    originalEndMinute = dragEndMinute,
-                    deltaMinutes = deltaMinutes,
-                )
-                .copy(date = if (commitDate) dragStartDate.plusDays(dayDelta.toLong()) else dragStartDate)
-            currentOnDraftChanged(nextDraft)
-        }
-
-        fun updateAutoScroll(direction: Float) {
-            autoScrollDirection = direction
-            if (direction == 0f) {
-                autoScrollJob?.cancel()
-                autoScrollJob = null
-                return
-            }
-            if (autoScrollJob?.isActive == true) return
-            autoScrollJob = launch {
-                while (true) {
-                    val consumed = timeScroll.dispatchRawDelta(autoScrollDirection * 9f)
-                    if (abs(consumed) < 0.5f) {
-                        delay(28)
-                    } else {
-                        totalDragPx += consumed
-                        publishDraft()
-                        delay(16)
-                    }
-                }
-            }
-        }
-
-        detectDragGestures(
-            onDragStart = {
-                onInteraction()
-                val snapshot = currentDraft
-                dragStartDate = snapshot.date
-                dragStartMinute = snapshot.start.minuteOfDay()
-                dragEndMinute = snapshot.end.minuteOfDay()
-                totalDragPx = 0f
-                dragX = 0f
-                draggingBody = mode == DraftDragMode.Move
-            },
-            onDrag = { change, dragAmount ->
-                change.consume()
-                totalDragPx += dragAmount.y
-                if (mode == DraftDragMode.Move) dragX += dragAmount.x
-                publishDraft()
-                val originalTopPx = ((dragStartMinute - visibleStartMinute) / 60f) * hourHeightPx
-                val visiblePointerY = originalTopPx + change.position.y + totalDragPx - timeScroll.value
-                updateAutoScroll(
-                    when {
-                        visiblePointerY < edgeThresholdPx -> -1f
-                        gridViewportHeightPx > 0 && visiblePointerY > gridViewportHeightPx - edgeThresholdPx -> 1f
-                        else -> 0f
-                    },
-                )
-            },
-            onDragEnd = {
-                updateAutoScroll(0f)
-                publishDraft(commitDate = true)
-                draggingBody = false
-                dragX = 0f
-            },
-            onDragCancel = {
-                updateAutoScroll(0f)
-                draggingBody = false
-                dragX = 0f
-            },
-        )
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .zIndex(80f)
-            .offset {
-                IntOffset(
-                    animatedDragX.roundToInt(),
-                    with(density) { (topDp - visualTopInsetDp).dp.roundToPx() },
-                )
-            }
-            .fillMaxWidth()
-            .height(touchHeightDp.dp)
-            .padding(horizontal = 2.dp),
-    ) {
-        Box(
-            modifier = Modifier
-                .matchParentSize()
-                .clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = null,
-                    onClick = onTap,
-                )
-                .testTag("editor-wireframe-body")
-                .draftDrag(DraftDragMode.Move),
-        )
-        Box(
-            modifier = Modifier
-                .offset(y = visualTopInsetDp.dp)
-                .fillMaxWidth()
-                .height(visualHeightDp.dp)
-                .clip(shape)
-                .background(wireframeColor.copy(alpha = 0.14f))
-                .border(1.5.dp, wireframeColor, shape)
-        )
-        // Start handle on the LEFT edge (top), end handle on the RIGHT edge (bottom) — diagonally
-        // opposite so they never collide even in a narrow 3-day column.
-        Box(
-            modifier = Modifier
-                .align(Alignment.TopStart)
-                .offset(x = (-6).dp, y = (visualTopInsetDp - 20f).dp)
-                .size(40.dp)
-                .testTag("editor-wireframe-start-handle")
-                .draftDrag(DraftDragMode.Start),
-            contentAlignment = Alignment.Center,
-        ) {
-            Box(
-                Modifier
-                    .size(16.dp)
-                    .clip(CircleShape)
-                    .background(wireframeColor)
-                    .border(2.dp, Color.White, CircleShape),
-            )
-        }
-        Box(
-            modifier = Modifier
-                .align(Alignment.TopEnd)
-                .offset(x = 6.dp, y = (visualTopInsetDp + visualHeightDp - 20f).dp)
-                .size(40.dp)
-                .testTag("editor-wireframe-end-handle")
-                .draftDrag(DraftDragMode.End),
-            contentAlignment = Alignment.Center,
-        ) {
-            Box(
-                Modifier
-                    .size(16.dp)
-                    .clip(CircleShape)
-                    .background(wireframeColor)
-                    .border(2.dp, Color.White, CircleShape),
-            )
-        }
-    }
-}
-
-@Composable
-private fun CurrentTimeLine(
-    now: LocalTime,
-    todayPageOffset: Dp,
-    dayWidth: Dp,
-    hourHeightDp: Float,
-    topOffset: Dp,
-    scrollOffset: Dp,
-    modifier: Modifier = Modifier,
-) {
-    if (now.hour !in DayStartHour..DayEndHour) return
-    val minuteOffset = (now.hour * 60 + now.minute) - DayStartHour * 60
-    val pillHeight = 18.dp
-    val lineCenterY = topOffset + (minuteOffset / 60f * hourHeightDp).dp - scrollOffset
-    val lineTopY = lineCenterY - (pillHeight / 2f)
-    if (lineTopY < topOffset) return
-    val darkPalette = MaterialTheme.colorScheme.background.isDark()
-    val indicatorColor = if (darkPalette) Color.White else Color(0xFF1D1511)
-    val pillBackground = if (darkPalette) Color.White.copy(alpha = 0.86f) else Color(0xFF1D1511).copy(alpha = 0.72f)
-    val pillTextColor = if (darkPalette) Color(0xFF151515) else Color.White
-    Box(modifier = modifier) {
-        Box(
-            modifier = Modifier
-                .offset(x = todayPageOffset)
-                .offset(y = lineTopY)
-                .width(TimeSidebarWidth)
-                .height(pillHeight)
-                .padding(horizontal = 2.dp)
-                .clip(RoundedCornerShape(9.dp))
-                .background(pillBackground),
-            contentAlignment = Alignment.Center,
-        ) {
-            Text(now.format(DateTimeFormatter.ofPattern("HH:mm")), color = pillTextColor, fontSize = 10.sp, lineHeight = 12.sp, fontWeight = FontWeight.Normal)
-        }
-        Box(
-            modifier = Modifier
-                .offset(x = TimeSidebarWidth + todayPageOffset, y = lineCenterY)
-                .height(1.dp)
-                .width(dayWidth)
-                .background(indicatorColor.copy(alpha = if (darkPalette) 0.62f else 0.52f)),
-        )
-    }
-}
-
-@Composable
-private fun DayTasksSheet(
-    day: LocalDate,
-    tasks: List<TaskEntity>,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onTaskClick: (TaskEntity) -> Unit,
-) {
-    val sheetScrollState = rememberScrollState()
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .verticalScroll(sheetScrollState)
-            .padding(horizontal = 22.dp, vertical = 18.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Text(
-            day.format(DateTimeFormatter.ofPattern("EEEE, d. MMMM", LocalAppLocale.current)),
-            color = WarmInk,
-            fontSize = 22.sp,
-            lineHeight = 26.sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-        Text(stringResource(R.string.task_count, tasks.size), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 14.sp, lineHeight = 17.sp)
-        tasks.forEach { task ->
-            TaskRow(
-                task = task,
-                taskColorMode = TaskColorMode.Collection,
-                onTaskStatusChanged = onTaskStatusChanged,
-                onClick = { onTaskClick(task) },
-            )
-        }
-        Spacer(Modifier.height(8.dp))
-    }
-}
-
-@Composable
-private fun AllDayChip(
-    title: String,
-    color: Color,
-    event: EventEntity? = null,
-    continuesFromPrevious: Boolean,
-    continuesToNext: Boolean,
-    onClick: () -> Unit,
-    morphDay: Boolean = false,
-) {
-    val darkPalette = CurrentDarkPalette
-    val visuals = remember(event?.status, event?.manualColor, event?.color, darkPalette) {
-        event?.cardVisuals(darkPalette = darkPalette)
-    }
-    val chipColor = visuals?.background ?: color
-    val textColor = visuals?.contentColor ?: if (chipColor.isDark()) Color.White else Color(0xFF1C1A18)
-    val shape = continuationShape(continuesFromPrevious, continuesToNext)
-    val eventTextStyle = tentativeReadableTextStyle(event?.isTentative() == true)
-    val pendingAlpha = event?.resourceHref?.let { pendingDeleteAlpha(it) } ?: 1f
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(24.dp)
-            // Option B: an all-day event morphs out of / into its month-cell pill (matched by uid),
-            // just like the timed cards, so the events already visible in the month rectangle travel
-            // into the all-day band instead of merely scaling with the grid.
-            .morphBounds(
-                key = "item-${event?.uid}",
-                enabled = EnablePerItemMorph && morphDay && event?.uid != null,
-                enter = MorphItemCardEnter,
-                exit = MorphItemCardExit,
-                boundsTransform = MorphItemBoundsTransform,
-            )
-            .alpha(pendingAlpha)
-            .drawContinuationBridge(chipColor, continuesFromPrevious, continuesToNext)
-            .background(chipColor, shape)
-            .then(
-                when {
-                    visuals?.dashedBorder == true && visuals.borderColor != null -> Modifier.dashedBorder(visuals.borderColor, 7.dp)
-                    visuals?.borderColor != null -> Modifier.border(1.dp, visuals.borderColor, shape)
-                    else -> Modifier
-                },
-            ),
-        contentAlignment = Alignment.CenterStart,
-    ) {
-        Box(
-            Modifier
-                .matchParentSize()
-                .clip(shape)
-                .clickable(onClick = onClick),
-        )
-        Text(
-            text = title,
-            color = textColor,
-            fontWeight = FontWeight.SemiBold,
-            fontSize = 12.sp,
-            lineHeight = 13.sp,
-            maxLines = 1,
-            softWrap = false,
-            overflow = TextOverflow.Clip,
-            textDecoration = visuals?.textDecoration,
-            style = eventTextStyle,
-            modifier = Modifier
-                .padding(horizontal = 6.dp, vertical = 3.dp)
-                .wrapContentWidth(unbounded = true),
-        )
-        event?.resourceHref?.let { href ->
-            PendingMutationBadge(href, Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp))
-        }
-    }
-}
-
-@Composable
-private fun AllDayTaskChip(
-    title: String,
-    color: Color,
-    completed: Boolean,
-    priority: Int?,
-    continuesFromPrevious: Boolean,
-    continuesToNext: Boolean,
-    status: String?,
-    resourceHref: String?,
-    onStatusChange: ((String) -> Unit)?,
-    onClick: () -> Unit,
-    morphUid: String? = null,
-) {
-    var lastCompleted by remember { mutableStateOf(completed) }
-    var burstKey by remember { mutableStateOf(0) }
-    LaunchedEffect(completed) {
-        if (completed && !lastCompleted) burstKey++
-        lastCompleted = completed
-    }
-    val alpha by animateFloatAsState(
-        targetValue = if (completed) 0.48f else 1f,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "allDayTaskAlpha",
-    )
-    val shape = continuationShape(continuesFromPrevious, continuesToNext)
-    val pendingAlpha = resourceHref?.let { pendingDeleteAlpha(it) } ?: 1f
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(24.dp)
-            // Option B: a single full-day task morphs out of / into its month-cell pill (matched by
-            // uid). Only when it's the lone all-day task — a collapsed "N tasks" chip has no
-            // single source pill to match.
-            .morphBounds(
-                key = "item-${morphUid}",
-                enabled = EnablePerItemMorph && morphUid != null,
-                enter = MorphItemCardEnter,
-                exit = MorphItemCardExit,
-                boundsTransform = MorphItemBoundsTransform,
-            )
-            .alpha(alpha * pendingAlpha)
-            .taskPriorityMotion(if (completed) null else priority, color)
-            .drawContinuationBridge(color, continuesFromPrevious, continuesToNext)
-            .background(color, shape),
-    ) {
-        TaskCardCompletionBurst(burstKey, color = if (color.isDark()) Color.White else WarmBrown, modifier = Modifier.matchParentSize())
-        Box(
-            Modifier
-                .matchParentSize()
-                .clip(shape)
-                .clickable(onClick = onClick),
-        )
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 5.dp, vertical = 3.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            val checkTint = if (color.isDark()) Color.White else Color(0xFF1C1A18)
-            if (status != null && onStatusChange != null) {
-                TaskStatusCheckbox(
-                    status = status,
-                    tint = checkTint,
-                    onStatusChange = onStatusChange,
-                    boxSize = 20.dp,
-                    iconSize = 16.dp,
-                )
-            } else {
-                Box(Modifier.size(20.dp), contentAlignment = Alignment.Center) {
-                    Icon(
-                        imageVector = if (completed) Icons.Default.CheckCircle else Icons.Default.RadioButtonUnchecked,
-                        contentDescription = null,
-                        tint = checkTint,
-                        modifier = Modifier.size(16.dp),
-                    )
-                }
-            }
-            Spacer(Modifier.width(4.dp))
-            Text(
-                text = title,
-                color = if (color.isDark()) Color.White else Color(0xFF1C1A18),
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 12.sp,
-                lineHeight = 13.sp,
-                maxLines = 1,
-                softWrap = false,
-                overflow = TextOverflow.Clip,
-                modifier = Modifier.wrapContentWidth(unbounded = true),
-            )
-        }
-        resourceHref?.let { href ->
-            PendingMutationBadge(href, Modifier.align(Alignment.TopEnd).offset(x = 2.dp, y = (-2).dp))
-        }
-    }
-}
 
 @Composable
 private fun CreateFabMenu(
@@ -8980,7 +5926,7 @@ private fun CompletionBurst(playKey: Int, color: Color, modifier: Modifier = Mod
 }
 
 @Composable
-private fun TaskCardCompletionBurst(playKey: Int, color: Color, modifier: Modifier = Modifier) {
+internal fun TaskCardCompletionBurst(playKey: Int, color: Color, modifier: Modifier = Modifier) {
     val progress = remember { Animatable(1f) }
     LaunchedEffect(playKey) {
         if (playKey > 0) {
@@ -15003,7 +11949,7 @@ private fun Modifier.bottomEdgeFade(color: Color, edgeHeight: Dp = 32.dp): Modif
 
 internal fun Modifier.editorInset(): Modifier = padding(horizontal = EditorHorizontalPadding)
 
-private fun Modifier.verticalClipAllowHorizontalOverflow(): Modifier = drawWithContent {
+internal fun Modifier.verticalClipAllowHorizontalOverflow(): Modifier = drawWithContent {
     clipRect(
         left = -size.width * 4f,
         top = 0f,
@@ -15204,14 +12150,14 @@ private fun TaskEntity.displayProgress(): Int =
         else -> 0
     }
 
-private fun taskPriorityIntensity(priority: Int?): Float {
+internal fun taskPriorityIntensity(priority: Int?): Float {
     val value = priority?.coerceIn(1, 9) ?: 9
     return ((9 - value) / 8f).coerceIn(0f, 1f)
 }
 
-private fun EventEntity.displayColor(): Int = manualColor ?: color
+internal fun EventEntity.displayColor(): Int = manualColor ?: color
 
-private data class EventCardVisuals(
+internal data class EventCardVisuals(
     val baseColor: Color,
     val background: Color,
     val contentColor: Color,
@@ -15220,7 +12166,7 @@ private data class EventCardVisuals(
     val textDecoration: TextDecoration?,
 )
 
-private fun EventEntity.cardVisuals(muted: Boolean = false, darkPalette: Boolean): EventCardVisuals {
+internal fun EventEntity.cardVisuals(muted: Boolean = false, darkPalette: Boolean): EventCardVisuals {
     val base = Color(displayColor())
     val tentative = isTentative()
     val cancelled = isCancelled()
@@ -15250,7 +12196,7 @@ private fun EventEntity.cardVisuals(muted: Boolean = false, darkPalette: Boolean
     )
 }
 
-private fun EventEntity.isTentative(): Boolean =
+internal fun EventEntity.isTentative(): Boolean =
     status.equals("TENTATIVE", ignoreCase = true)
 
 private fun EventEntity.isCancelled(): Boolean =
@@ -15518,7 +12464,7 @@ private fun pendingMutationFor(resourceHref: String): PendingMutationEntity? {
 }
 
 @Composable
-private fun pendingDeleteAlpha(resourceHref: String): Float {
+internal fun pendingDeleteAlpha(resourceHref: String): Float {
     val pending = pendingMutationFor(resourceHref)
     val exiting = resourceHref in LocalExitingResourceHrefs.current
     val target = when {
@@ -15535,7 +12481,7 @@ private fun pendingDeleteAlpha(resourceHref: String): Float {
 }
 
 @Composable
-private fun PendingMutationBadge(
+internal fun PendingMutationBadge(
     resourceHref: String,
     modifier: Modifier = Modifier,
 ) {
@@ -15605,7 +12551,7 @@ internal fun String.localizedParticipantRoleLabel(): String =
     ParticipantRoleOption.entries.firstOrNull { it.value.equals(this, ignoreCase = true) }?.localizedLabel() ?: this
 
 @Composable
-private fun EventParticipantStack(
+internal fun EventParticipantStack(
     attendees: List<CalendarParticipant>,
     eventColor: Color,
     contentColor: Color,
@@ -15682,7 +12628,7 @@ private fun EventParticipantStack(
 }
 
 @Composable
-private fun CompactParticipantStatusDots(
+internal fun CompactParticipantStatusDots(
     attendees: List<CalendarParticipant>,
     modifier: Modifier = Modifier,
     maxVisible: Int = 3,
@@ -15952,7 +12898,7 @@ private fun taskStatusOptionLabel(value: String): String = when (value) {
 }
 
 /** Normalised effective status, falling back to the legacy isCompleted boolean. */
-private fun TaskEntity.effectiveStatus(): String = status?.uppercase()
+internal fun TaskEntity.effectiveStatus(): String = status?.uppercase()
     ?: if (isCompleted) "COMPLETED" else "NEEDS-ACTION"
 
 /**
@@ -16351,7 +13297,7 @@ private fun CollectionEntity.localizedKindLabel(): String = buildList {
 private fun CollectionEntity.isVisibleInSettingsCalendarList(): Boolean =
     !(href.isLocalCollectionHrefUi() && !isEnabled)
 
-private fun TaskEntity.displayColor(mode: TaskColorMode): Int =
+internal fun TaskEntity.displayColor(mode: TaskColorMode): Int =
     manualColor ?: when (mode) {
         TaskColorMode.Collection -> color
         TaskColorMode.Priority -> priority?.let { priorityColor(it).toArgb() } ?: color
@@ -16381,7 +13327,7 @@ private fun String.locationDisplayName(): String =
         .substringBefore(',')
         .trim()
 
-private fun String.cardLocationText(mapVerified: Boolean?): String {
+internal fun String.cardLocationText(mapVerified: Boolean?): String {
     val cleaned = cleanCalendarDisplayText()
     if (cleaned.isBlank()) return ""
     return if (mapVerified == true) cleaned.substringBefore(',').trim() else cleaned
@@ -16485,7 +13431,7 @@ private fun CollectionEntity.typeLabel(): String = when {
     else -> "Events"
 }
 
-private fun continuationShape(continuesFromPrevious: Boolean, continuesToNext: Boolean): RoundedCornerShape =
+internal fun continuationShape(continuesFromPrevious: Boolean, continuesToNext: Boolean): RoundedCornerShape =
     RoundedCornerShape(
         topStart = if (continuesFromPrevious) 0.dp else 8.dp,
         bottomStart = if (continuesFromPrevious) 0.dp else 8.dp,
@@ -16493,7 +13439,7 @@ private fun continuationShape(continuesFromPrevious: Boolean, continuesToNext: B
         bottomEnd = if (continuesToNext) 0.dp else 8.dp,
     )
 
-private fun Modifier.drawContinuationBridge(
+internal fun Modifier.drawContinuationBridge(
     color: Color,
     continuesFromPrevious: Boolean,
     continuesToNext: Boolean,
@@ -16533,7 +13479,7 @@ private fun LocalTime.nextDraftStart(): LocalTime {
 private fun LocalTime.defaultDraftEnd(durationMinutes: Int = 60): LocalTime =
     (minuteOfDay() + durationMinutes.coerceIn(15, 24 * 60)).coerceAtMost((DayEndHour + 1) * 60 - 1).toDraftLocalTime()
 
-private fun LocalTime.minuteOfDay(): Int = hour * 60 + minute
+internal fun LocalTime.minuteOfDay(): Int = hour * 60 + minute
 
 private fun LocalDate.startOfDayMillis(): Long =
     atStartOfDay(ZoneId.systemDefault()).toInstant().toEpochMilli()
@@ -16581,15 +13527,15 @@ internal fun taskDateTimeRangeInvalid(
     return !endDate.atTime(endTime).isAfter(startDate.atTime(startTime))
 }
 
-private fun Int.toDraftLocalTime(): LocalTime {
+internal fun Int.toDraftLocalTime(): LocalTime {
     val minute = coerceIn(DayStartHour * 60, (DayEndHour + 1) * 60 - 1)
     return LocalTime.of(minute / 60, minute % 60)
 }
 
-private fun Int.snapDraftMinute(): Int =
+internal fun Int.snapDraftMinute(): Int =
     ((this + DraftSnapMinutes / 2) / DraftSnapMinutes) * DraftSnapMinutes
 
-private fun EditorSchedulePreview.withDraggedMinutes(
+internal fun EditorSchedulePreview.withDraggedMinutes(
     mode: DraftDragMode,
     originalStartMinute: Int,
     originalEndMinute: Int,
@@ -16625,7 +13571,7 @@ private fun EditorSchedulePreview.withDraggedMinutes(
     return copy(start = startMinute.toDraftLocalTime(), end = endMinute.toDraftLocalTime())
 }
 
-private fun List<LocalDate>.allDayAreaHeight(
+internal fun List<LocalDate>.allDayAreaHeight(
     events: List<EventEntity>,
     tasks: List<TaskEntity>,
     maxVisibleItems: Int,
@@ -16674,13 +13620,13 @@ private fun List<LocalDate>.allDayAreaHeight(
     return if (displayedRowsWithDraft == 0) 22.dp else (displayedRowsWithDraft * 24 + (displayedRowsWithDraft - 1) * 5 + 15).dp
 }
 
-private fun firstFreeLaneIndex(occupiedLanes: Set<Int>): Int {
+internal fun firstFreeLaneIndex(occupiedLanes: Set<Int>): Int {
     var lane = 0
     while (lane in occupiedLanes) lane++
     return lane
 }
 
-private fun List<LocalDate>.hasAllDayOverflow(
+internal fun List<LocalDate>.hasAllDayOverflow(
     events: List<EventEntity>,
     tasks: List<TaskEntity>,
     maxVisibleItems: Int,
@@ -16694,7 +13640,7 @@ private fun List<LocalDate>.hasAllDayOverflow(
     }
 }
 
-private fun List<EventEntity>.indexEventsByDay(): Map<LocalDate, List<EventEntity>> {
+internal fun List<EventEntity>.indexEventsByDay(): Map<LocalDate, List<EventEntity>> {
     val result = linkedMapOf<LocalDate, MutableList<EventEntity>>()
     forEach { event ->
         val start = event.startsAtMillis.toDate()
@@ -16719,7 +13665,7 @@ private fun EventEntity.smoothRemovalKey(): String =
 private fun TaskEntity.smoothRemovalKey(): String =
     "${resourceHref}:${startAtMillis ?: dueAtMillis ?: completedAtMillis ?: 0L}"
 
-private fun List<TaskEntity>.indexTasksByDay(): Map<LocalDate, List<TaskEntity>> =
+internal fun List<TaskEntity>.indexTasksByDay(): Map<LocalDate, List<TaskEntity>> =
     flatMap { task -> task.visibleDates().map { it to task } }
         .groupBy({ it.first }, { it.second })
 
@@ -16810,7 +13756,7 @@ private sealed interface RecurringSaveRequest {
     data class Task(val task: TaskEntity, val payload: TaskEditPayload) : RecurringSaveRequest
 }
 
-private enum class SheetSnap {
+internal enum class SheetSnap {
     Expanded,
     Half,
     Quarter,
@@ -16818,7 +13764,7 @@ private enum class SheetSnap {
     EditorTiny,
 }
 
-private enum class SheetAnchorMode {
+internal enum class SheetAnchorMode {
     ContentFit,
     Editor,
 }
@@ -16946,7 +13892,7 @@ private data class MonthDayMarkers(
     val hasMore: Boolean,
 )
 
-private enum class DraftDragMode {
+internal enum class DraftDragMode {
     Move,
     Start,
     End,
@@ -17009,7 +13955,7 @@ internal fun EditorTransferDraft.withDestinationReminderDefaults(destinationDefa
     )
 }
 
-private fun buildAllDayOverlayItems(
+internal fun buildAllDayOverlayItems(
     events: List<EventEntity>,
     tasks: List<TaskEntity>,
     taskColorMode: TaskColorMode,
@@ -17161,7 +14107,7 @@ internal fun Color.isDark(): Boolean {
     return luminance < 0.55f
 }
 
-private fun Color.blendWith(target: Color, amount: Float): Color {
+internal fun Color.blendWith(target: Color, amount: Float): Color {
     val t = amount.coerceIn(0f, 1f)
     return Color(
         red = red + (target.red - red) * t,
